% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-tools.R
\name{addClstrInf}
\alias{addClstrInf}
\title{Add taxonomic and sequence info to clusters}
\usage{
addClstrInf(clstrs, phylt_nds, txid, sqs, typ)
}
\arguments{
\item{clstrs}{Clusters}

\item{phylt_nds}{PhyLoTa data.frame}

\item{txid}{Taxonomic node ID}

\item{sqs}{Sequneces}

\item{typ}{Subtree of direct?}
}
\description{
input: List of clstrs
}
