% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage2-tools.R
\name{agmntSqRcrds}
\alias{agmntSqRcrds}
\alias{seqrec_augment}
\title{Augment sequence records list}
\usage{
seqrec_augment(sqs, txdct)
}
\arguments{
\item{sqs}{List of SqRcrds}

\item{txdct}{Taxonomic Dictionary}
}
\description{
Convert to sqsrcrds and add taxids
}
