\name{aotus}
\alias{aotus}
\docType{data}
\title{Example phylota object (aotus)}
\description{
  Final phylota object produced by running pipeline for the night monkeys, Aotus (9504).
}
\usage{
   data(aotus)
}
\format{
  \code{aotus} is generated from read_phylota().
}
\examples{
data(aotus) # load object
str(aotus)
}
\keyword{datasets}