% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes.R
\docType{class}
\name{ArchiveCluster-class}
\alias{ArchiveCluster-class}
\alias{ArchiveCluster-method}
\alias{as.character,ArchiveCluster-method}
\alias{show,ArchiveCluster-method}
\alias{print,ArchiveCluster-method}
\alias{str,ArchiveCluster-method}
\alias{summary,ArchiveCluster-method}
\alias{[[,ArchiveCluster,character-method}
\alias{[,ArchiveCluster,character,missing,missing-method}
\title{ArchiveCluster-class}
\usage{
\S4method{as.character}{ArchiveCluster}(x)

\S4method{show}{ArchiveCluster}(object)

\S4method{print}{ArchiveCluster}(x)

\S4method{str}{ArchiveCluster}(object, max.level = 2L, ...)

\S4method{summary}{ArchiveCluster}(object)

\S4method{[[}{ArchiveCluster,character}(x, i)

\S4method{[}{ArchiveCluster,character,missing,missing}(x, i, j, ...,
  drop = TRUE)
}
\arguments{
\item{x}{\code{ArchiveCluster} object}

\item{object}{\code{ArchiveCluster} object}

\item{max.level}{Maximum level of nesting for str()}

\item{...}{Further arguments for str()}

\item{i}{cid(s)}

\item{j}{Unused}

\item{drop}{Unused}
}
\description{
Multiple cluster records.
}
\section{Slots}{

\describe{
\item{\code{ids}}{Vector of Cluster Record IDs}

\item{\code{cls}}{List of ArchiveCluster named by ID}
}}

