% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes.R
\docType{class}
\name{ArchiveSequence-class}
\alias{ArchiveSequence-class}
\alias{ArchiveSequence-method}
\alias{as.character,ArchiveSequence-method}
\alias{show,ArchiveSequence-method}
\alias{print,ArchiveSequence-method}
\alias{str,ArchiveSequence-method}
\alias{summary,ArchiveSequence-method}
\alias{[[,ArchiveSequence,character-method}
\alias{[,ArchiveSequence,character,missing,missing-method}
\title{ArchiveSequence-class}
\usage{
\S4method{as.character}{ArchiveSequence}(x)

\S4method{show}{ArchiveSequence}(object)

\S4method{print}{ArchiveSequence}(x)

\S4method{str}{ArchiveSequence}(object, max.level = 2L, ...)

\S4method{summary}{ArchiveSequence}(object)

\S4method{[[}{ArchiveSequence,character}(x, i)

\S4method{[}{ArchiveSequence,character,missing,missing}(x, i, j, ...,
  drop = TRUE)
}
\arguments{
\item{x}{\code{ArchiveSequence} object}

\item{object}{\code{ArchiveSequence} object}

\item{max.level}{Maximum level of nesting for str()}

\item{...}{Further arguments for str()}

\item{i}{sid(s)}

\item{j}{Unused}

\item{drop}{Unused}
}
\description{
Multiple sequence records containing sequence data.
}
\details{
Sequences are stored as raw. Use rawToChar().
}
\section{Slots}{

\describe{
\item{\code{ids}}{Vector of Sequence Record IDs}

\item{\code{nncltds}}{Vector of sequence lengths}

\item{\code{nambgs}}{Vector of number of ambiguous nucleotides}

\item{\code{txids}}{Vector source txid associated with each sequence}

\item{\code{sqs}}{List of RecordSequences named by ID}
}}

