% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{blast.all.vs.all}
\alias{blast.all.vs.all}
\title{BLAST all vs all}
\usage{
blast.all.vs.all(dbname = "blastdb.fa", evalue.cutoff = 1e-10,
  outfile = "blastout.txt", dir = ".", prmtrs = prmtrs)
}
\description{
TODO
}
\examples{
# TODO
}
