% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline-tools.R
\name{blast_setup}
\alias{blast_setup}
\title{Ensures NCBI BLAST tools are installed}
\usage{
blast_setup(d, v, wd)
}
\arguments{
\item{d}{Directory to NCBI BLAST tools}

\item{v}{v, T/F}

\item{wd}{Working directory}
}
\description{
Ensures NCBI BLAST executables are installed on the 
system. Tests version number of BLAST tools.
}
\details{
BLAST tools must be version >=2.0
}
\seealso{
\link{parameters_setup}
}
