\name{blst_rs}
\alias{blst_rs}
\docType{data}
\title{Example BLAST Results}
\description{
  Simple BLAST results for testing and examples.
}
\usage{
   data(blst_rs)
}
\format{
  \code{blst_rs} is a produced by a blast function.
}
\examples{
data(blst_rs) # load object
summary(blst_rs)
}
\keyword{datasets}