% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast-tools.R
\name{blstN}
\alias{blstN}
\title{BLAST all vs all}
\usage{
blstN(dbfl, outfl, wd, eval_ctoff = 1e-10, verbose = FALSE)
}
\arguments{
\item{dbfl}{Database file}

\item{outfl}{Output file}

\item{wd}{Working directory}

\item{eval_ctoff}{Minimum E-value}

\item{verbose}{Verbose? T/F}
}
\description{
Use \code{blastn} to BLAST all vs all using
a BLAST database
}
