% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-tools.R
\name{blstSqs}
\alias{blstSqs}
\title{BLAST All vs All}
\usage{
blstSqs(txid, typ, sqs, wd, verbose = FALSE)
}
\arguments{
\item{txid}{Taxonomic node ID, numeric}

\item{typ}{Cluster type, 'direct' or 'subtree'}

\item{sqs}{Sequences}

\item{wd}{Working directory}

\item{verbose}{Verbose? T/F}
}
\description{
Return blast results from
BLASTing all vs all for given sequences
}
