\name{bromeliads}
\alias{bromeliads}
\docType{data}
\title{Example phylota object (bromeliads)}
\description{
  Final phylota object produced by running pipeline for the bromeliads, Bromeliaceae (4613).
}
\usage{
   data(bromeliads)
}
\format{
  \code{bromeliads} is generated from read_phylota().
}
\examples{
data(bromeliads) # load object
str(bromeliads)
}
\keyword{datasets}