% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrez-tools.R
\name{btchDwnld}
\alias{btchDwnld}
\title{btchDwnld}
\usage{
btchDwnld(txid, ps, drct = FALSE, lvl = 0)
}
\arguments{
\item{txid}{NCBI taxonomic ID}

\item{drct}{Node-level only or subtree as well? Default FALSE.}
}
\value{
Vector of sequence records
}
\description{
Downloads sequences from GenBank in 500 ID batches.
}
