% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylota-tools.R
\name{byLineage}
\alias{byLineage}
\title{Find clusters with a given taxonomic group}
\usage{
byLineage(clstrs_obj, nm)
}
\arguments{
\item{clstrs_obj}{clstrs_obj}

\item{nm}{Taxonomic group name}
}
\description{
Returns a boolean vector of all the
clusters that have representatives of a given
taxonomic group.
}
