% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-tools-phylota-calc.R
\name{calc_mad}
\alias{calc_mad}
\title{Calculate MAD score}
\usage{
calc_mad(phylota, cid)
}
\arguments{
\item{phylota}{Phylota object}

\item{cid}{CLuster ID(s)}
}
\value{
vector
}
\description{
For all sequences in a cluster(s) the
MAD score.
}
\details{
MAD is a measure of the deviation in 
sequence length of a cluster. Values range from 0 to
1. Clusters with values close to 1 have sequences with
similar lengths.
}
