% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-tools-phylota-calc.R
\name{calc_wrdfrq}
\alias{calc_wrdfrq}
\title{Calculate word frequencies}
\usage{
calc_wrdfrq(phylota, cid, min_frq = 0.1, type = c("dfln", "nm"),
  ignr_pttrn = "[^a-z0-9]")
}
\arguments{
\item{phylota}{Phylota object}

\item{cid}{CLuster ID(s)}

\item{min_frq}{Minimum frequency}

\item{type}{Definitions (dfln) or features (nm)}

\item{ignr_pttrn}{Ignore pattern, REGEX for text to ignore.}
}
\value{
list
}
\description{
For all sequences in a cluster(s)
calculate the frequency of separate words in
either the sequnece definitions or the reported
feature name.
}
\details{
By default, anything that is not alphanumeric is 
ignored.
'dfln' and 'nm' match the slot names in a SqRcrd, see
list_sqrcrd_slots().
}
