% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-tools.R
\name{calcClstrs}
\alias{calcClstrs}
\title{Calculate clusters for all sequences in WD}
\usage{
calcClstrs(wd, txid, phylt_nds, verbose = FALSE)
}
\arguments{
\item{phylt_nds}{PhyLoTa nodes data.frame}
}
\description{
TODO
}
