% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.R
\name{calcFrPrp}
\alias{calcFrPrp}
\title{Calculate evolutionary distinctness}
\usage{
calcFrPrp(tree, tids, progress = "none")
}
\arguments{
\item{tree}{\code{TreeMan} object}

\item{tids}{tip IDs}

\item{progress}{name of the progress bar to use, see \code{\link[plyr]{create_progress_bar}}}
}
\description{
Returns the evolutationary distinctness of ids using the fair proportion metric.
}
\details{
The fair proportion metric calculates the evolutionary distinctness of tips
in a tree through summing the total amount of branch length each tip represents, where
each branch in the tree is evenly divided between all descendants. Parallelizable.
}
\examples{

tree <- randTree(10)
calcFrPrp(tree, tree["tips"])
}
\references{
Isaac, N.J.B., Turvey, S.T., Collen, B., Waterman, C. and Baillie, J.E.M. (2007).
 Mammals on the EDGE: conservation priorities based on threat and phylogeny. PLoS ONE, 2, e296.
}
\seealso{
\code{\link{calcPhyDv}}, \code{\link{calcPrtFrPrp}},
\url{https://github.com/DomBennett/treeman/wiki/calc-methods}
}
