% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-cache.R
\name{chckGIs}
\alias{chckGIs}
\title{Check if GIs exist}
\usage{
chckGIs(wd, txid)
}
\arguments{
\item{wd}{Working directory}

\item{txid}{Taxonomic ID, numeric}
}
\description{
Check if GIs are already
downloaded for a txid
}
