% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-api.R
\name{chckSrchObj}
\alias{chckSrchObj}
\title{Check an object returned from rentrez}
\usage{
chckSrchObj(obj)
}
\arguments{
\item{obj}{Object returned from rentrez function}
}
\description{
Returns T/F. Checks if object
returned from rentrez function is as expected.
}
