% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{chckStgs}
\alias{chckStgs}
\title{Check to and frm stage numbers}
\usage{
chckStgs(frm, to)
}
\arguments{
\item{frm}{Starting stage}

\item{to}{Ending stage}
}
\description{
Ensure stage numbers, return stage print message.
}
\seealso{
\code{\link{run}}
}
