% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-cache.R
\name{chkObj}
\alias{chkObj}
\alias{obj_check}
\title{Check if an object exists}
\usage{
obj_check(wd, nm)
}
\arguments{
\item{wd}{Working directory}

\item{nm}{Object name}
}
\description{
Check if an object exists in the cache.
}
