% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage2.R
\name{cldIdntfy}
\alias{cldIdntfy}
\alias{clade_select}
\title{Get all node IDs that will be processed}
\usage{
clade_select(txdct, ps)
}
\arguments{
\item{txdct}{TxDct}

\item{ps}{Parameters}
}
\description{
All nodes with less than maximum number
of nodes and sequences.
}
