% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes.R
\docType{class}
\name{ClRcrd-class}
\alias{ClRcrd-class}
\alias{ClRcrd-method}
\alias{as.character,ClRcrd-method}
\alias{show,ClRcrd-method}
\alias{print,ClRcrd-method}
\alias{str,ClRcrd-method}
\alias{summary,ClRcrd-method}
\title{ClRcrd-class}
\usage{
\S4method{as.character}{ClRcrd}(x)

\S4method{show}{ClRcrd}(object)

\S4method{print}{ClRcrd}(x)

\S4method{str}{ClRcrd}(object, max.level = 2L, ...)

\S4method{summary}{ClRcrd}(object)
}
\arguments{
\item{x}{\code{ClRcrd} object}

\item{object}{\code{ClRcrd} object}

\item{max.level}{Maximum level of nesting for str()}

\item{...}{Further arguments for str()}
}
\description{
Cluster record contains all information on a cluster.
}
\section{Slots}{

\describe{
\item{\code{id}}{Cluster ID, integer}

\item{\code{sids}}{Sequence IDs}

\item{\code{nsqs}}{Number of sequences}

\item{\code{txids}}{Source txids for sequences}

\item{\code{ntx}}{Number of taxa}

\item{\code{typ}}{Cluster type: direct, subtree or merged}

\item{\code{seed}}{Seed sequence ID}

\item{\code{prnt}}{Parent taxonomic ID}
}}

