% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-clstr.R
\docType{class}
\name{Clstr-class}
\alias{Clstr-class}
\alias{Clstr-method}
\alias{as.character,Clstr-method}
\alias{show,Clstr-method}
\alias{print,Clstr-method}
\alias{str,Clstr-method}
\alias{summary,Clstr-method}
\title{Clstr-class}
\usage{
\S4method{as.character}{Clstr}(x)

\S4method{show}{Clstr}(object)

\S4method{print}{Clstr}(x)

\S4method{str}{Clstr}(object, max.level = 2L, ...)

\S4method{summary}{Clstr}(object)
}
\arguments{
\item{x}{\code{Clstr} object}

\item{object}{\code{Clstr} object}
}
\description{
Cluster record contains all information on a cluster.
}
\section{Slots}{

\describe{
\item{\code{id}}{Cluster ID, integer}

\item{\code{sids}}{Sequence IDs}

\item{\code{txids}}{Source txids for sequences}

\item{\code{typ}}{Cluster type: direct, subtree or merged}

\item{\code{seed}}{Seed sequence ID}

\item{\code{prnt}}{Parent taxonomic ID}
}}

\seealso{
\code{\link{genClstr}}
}
