% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-tools.R
\name{clstrBlstRs}
\alias{clstrBlstRs}
\title{Cluster BLAST Results}
\usage{
clstrBlstRs(blst_rs, infrmtv = TRUE)
}
\arguments{
\item{blst_rs}{BLAST results}

\item{infrmtv}{Informative? T/F}
}
\description{
TODO
}
