% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage4.R
\name{clstrClstrs}
\alias{clstrClstrs}
\title{Cluster sets of clusters identified in cluster stage}
\usage{
clstrClstrs(ps)
}
\arguments{
\item{ps}{Parameters}
}
\description{
Loads cluster sets from cache. Extracts seed sequences
and runs all-v-all BLAST of seeds to identify sister clusters.
Sisters are then merged. An object of all sequences and clusters
is then saved in cache.
}
