% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unused.R
\name{clstrPhylt}
\alias{clstrPhylt}
\title{Reformat cluster obj for PhyLoTa table}
\usage{
clstrPhylt(clstrs)
}
\arguments{
\item{clstrs}{Clusters}
}
\description{
Returns a data frame with columns matchine the fields in
PhyLoTA's cluster table
}
