% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl.R
\name{Cluster}
\alias{Cluster}
\alias{cluster}
\title{Cluster}
\usage{
cluster(taxon, nodes, seqs, blast.results = NULL, direct = FALSE,
  informative = FALSE, prmtrs = prmtrs)
}
\description{
TODO
}
\examples{
# TODO
}
