% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage3-tools.R
\name{clusterAll}
\alias{clusterAll}
\alias{cluster_all}
\title{Hierarchically cluster all sequences of a txid}
\usage{
cluster_all(txid, sqs, txdct, ps, lvl = 0)
}
\arguments{
\item{txid}{Taxonomic ID}

\item{sqs}{Sequence object of all downloaded sequences}

\item{txdct}{PhyLoTa table}

\item{ps}{Parameters}

\item{lvl}{Log level}
}
\description{
Identifies all direct and subtree clusters
for a taxonomic ID.
}
