% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes.R
\docType{class}
\name{ClusterArc-class}
\alias{ClusterArc-class}
\alias{ClusterArc-method}
\alias{as.character,ClusterArc-method}
\alias{show,ClusterArc-method}
\alias{print,ClusterArc-method}
\alias{str,ClusterArc-method}
\alias{summary,ClusterArc-method}
\alias{[[,ClusterArc,character-method}
\alias{[,ClusterArc,character,missing,missing-method}
\title{ClusterArc-class}
\usage{
\S4method{as.character}{ClusterArc}(x)

\S4method{show}{ClusterArc}(object)

\S4method{print}{ClusterArc}(x)

\S4method{str}{ClusterArc}(object, max.level = 2L, ...)

\S4method{summary}{ClusterArc}(object)

\S4method{[[}{ClusterArc,character}(x, i)

\S4method{[}{ClusterArc,character,missing,missing}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{\code{ClusterArc} object}

\item{object}{\code{ClusterArc} object}

\item{max.level}{Maximum level of nesting for str()}

\item{...}{Further arguments for str()}

\item{i}{cid(s)}

\item{j}{Unused}

\item{drop}{Unused}
}
\description{
Multiple cluster records.
}
\section{Slots}{

\describe{
\item{\code{ids}}{Vector of Cluster Record IDs}

\item{\code{cls}}{List of ClusterArc named by ID}
}}

