% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage3-tools.R
\name{clusterBlstRs}
\alias{clusterBlstRs}
\alias{blast_cluster}
\title{Cluster BLAST Results}
\usage{
blast_cluster(blast_res)
}
\arguments{
\item{blst_rs}{BLAST results}
}
\value{
List of list
}
\description{
Find single-linkage clusters from
BLAST results. Identifies seed sequence.
}
