% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes.R
\docType{class}
\name{ClusterRec-class}
\alias{ClusterRec-class}
\alias{ClusterRec-method}
\alias{as.character,ClusterRec-method}
\alias{show,ClusterRec-method}
\alias{print,ClusterRec-method}
\alias{str,ClusterRec-method}
\alias{summary,ClusterRec-method}
\title{ClusterRec-class}
\usage{
\S4method{as.character}{ClusterRec}(x)

\S4method{show}{ClusterRec}(object)

\S4method{print}{ClusterRec}(x)

\S4method{str}{ClusterRec}(object, max.level = 2L, ...)

\S4method{summary}{ClusterRec}(object)
}
\arguments{
\item{x}{\code{ClusterRec} object}

\item{object}{\code{ClusterRec} object}

\item{max.level}{Maximum level of nesting for str()}

\item{...}{Further arguments for str()}
}
\description{
Cluster record contains all information on a cluster.
}
\section{Slots}{

\describe{
\item{\code{id}}{Cluster ID, integer}

\item{\code{sids}}{Sequence IDs}

\item{\code{nsqs}}{Number of sequences}

\item{\code{txids}}{Source txids for sequences}

\item{\code{ntx}}{Number of taxa}

\item{\code{typ}}{Cluster type: direct, subtree or merged}

\item{\code{seed}}{Seed sequence ID}

\item{\code{prnt}}{Parent taxonomic ID}
}}

