% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl.R
\name{clusters.ci_gi.seqs.create}
\alias{clusters.ci_gi.seqs.create}
\title{Clusters CI GI Sequences}
\usage{
clusters.ci_gi.seqs.create(root.taxon, nodesfile, files = list(clusters =
  "clusters.tsv", ci_gi = "ci_gi.tsv", seqs = "seqs.tsv"),
  informative = FALSE, prmtrs = prmtrs)
}
\description{
TODO
}
\examples{
# TODO
}
