% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage3.R
\name{clusters_run}
\alias{clusters_run}
\title{Run the cluster stage}
\usage{
clusters_run(wd)
}
\arguments{
\item{wd}{Working directory}
}
\description{
Run the third stage of the phylotaR pipeline, cluster.
This stage hierarchically traverses the taxonomy identifying all
direct and subtree clusters from downloaded sequences. Any
taxonomic nodes too small for cluster identification are placed
into paraphyletic clusters.
}
\examples{
\dontrun{
  
  # individually run stages
  setup(wd = 'aotus', txid = 9504)
  taxise_run(wd = 'aotus')
  download_run(wd = 'aotus')
  clusters_run(wd = 'aotus')
}
}
\seealso{
Other run-public: \code{\link{ClstrArc-class}},
  \code{\link{ClstrRec-class}},
  \code{\link{Phylota-class}}, \code{\link{SeqArc-class}},
  \code{\link{SeqRec-class}}, \code{\link{TaxDict-class}},
  \code{\link{TaxRec-class}}, \code{\link{clusters2_run}},
  \code{\link{parameters_reset}}, \code{\link{reset}},
  \code{\link{restart}}, \code{\link{run}},
  \code{\link{setup}}, \code{\link{taxise_run}}
}
