% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage3-tools.R
\name{clusterSqs}
\alias{clusterSqs}
\alias{cluster_sqs}
\title{Identify clusters from sequences}
\usage{
cluster_sqs(txid, sqs, ps, lvl, typ = c("direct", "subtree", "paraphyly"))
}
\arguments{
\item{txid}{Taxonomic ID}

\item{sqs}{Sequence object of sequences to be BLASTed}

\item{ps}{Parameters}

\item{lvl}{Log level}

\item{typ}{Direct or Subtree?}
}
\description{
Given a sequence object, this function will generate
a list of cluster objects using BLAST
}
