% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-system.R
\name{cmdLn}
\alias{cmdLn}
\alias{cmdln}
\title{Run a command via terminal/command prompt}
\usage{
cmdln(cmd, args, lgfl = NULL)
}
\arguments{
\item{cmd}{Command to be run}

\item{args}{Vector of command arguments, each parameter and value must
be a separate element}

\item{lgfl}{File to which stdout/err will be written}
}
\description{
Provide the command and arguments as a vector.
Also can take a lgfl to which all stdout and stderr is written.
If lgfl is not provided, a list is returned of 'status', 'stdout'
and 'stderr'. Else only the status is returned - 1 success, 0 failed.
}
\details{
Note, stdout/err are returned as 'raw'. Use rawToChar() to
convert to characters.
}
