% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.R
\name{cTrees}
\alias{cTrees}
\title{cTrees}
\usage{
cTrees(x, ...)
}
\arguments{
\item{x}{\code{TreeMan} or \code{TreeMen} objects}

\item{...}{more \code{TreeMan} or \code{TreeMen} objects}
}
\description{
Return \code{TreeMen} of concatenated trees.
}
\details{
Concatenate trees into single \code{TreeMen} object.
}
\examples{

trees <- cTrees(randTree(10), randTree(10))
}
\seealso{
\code{\link{TreeMen-class}}, \code{\link{TreeMan-class}}, \code{\link{list-to-TreeMen}}
}
