\name{cycads}
\alias{cycads}
\docType{data}
\title{Example phylota object (cycads)}
\description{
  Final phylota object produced by running pipeline for the cycads, Cycadidae (1445963).
}
\usage{
   data(cycads)
}
\format{
  \code{cycads} is generated from read_phylota().
}
\examples{
data(cycads) # load object
str(cycads)
}
\keyword{datasets}