% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes.R
\docType{class}
\name{DictionaryTaxon-class}
\alias{DictionaryTaxon-class}
\alias{DictionaryTaxon-method}
\alias{as.character,DictionaryTaxon-method}
\alias{show,DictionaryTaxon-method}
\alias{print,DictionaryTaxon-method}
\alias{str,DictionaryTaxon-method}
\alias{summary,DictionaryTaxon-method}
\title{DictionaryTaxon-class}
\usage{
\S4method{as.character}{DictionaryTaxon}(x)

\S4method{show}{DictionaryTaxon}(object)

\S4method{print}{DictionaryTaxon}(x)

\S4method{str}{DictionaryTaxon}(object, max.level = 2L, ...)

\S4method{summary}{DictionaryTaxon}(object)
}
\arguments{
\item{x}{\code{DictionaryTaxon} object}

\item{object}{\code{DictionaryTaxon} object}

\item{max.level}{Maximum level of nesting for str()}

\item{...}{Further arguments for str()}
}
\description{
Taxonomic dictionary contains a taxonomic
tree and NCBI taxonomy data for all taxonomic IDs.
}
\section{Slots}{

\describe{
\item{\code{txids}}{Taxonomic IDs of taxon records}

\item{\code{rcrds}}{Environment of records}

\item{\code{prnt}}{Parent taxonomic ID}

\item{\code{txtr}}{Taxonomic tree}

\item{\code{indx}}{Taxonomic ID index for tree IDs}
}}

