% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage2.R
\name{download_run}
\alias{download_run}
\title{Run download stage}
\usage{
download_run(wd)
}
\arguments{
\item{wd}{Working directory}
}
\description{
Run the second stage of phylotaR, download. This stage
downloads sequences for all nodes with sequence numbers less than
mxsqs. It hierarchically traverses the taxonomy for each node and
downloads direct and subtree sequences for all descendants.
}
