\name{dragonflies}
\alias{dragonflies}
\docType{data}
\title{Example phylota object (dragonflies)}
\description{
  Final phylota object produced by running pipeline for the dragonflies, Anisoptera (6962).
}
\usage{
   data(dragonflies)
}
\format{
  \code{dragonflies} is generated from read_phylota().
}
\examples{
data(dragonflies) # load object
str(dragonflies)
}
\keyword{datasets}