% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-drop.R
\name{drop_cls}
\alias{drop_cls}
\title{Drop cluster records from phylota object}
\usage{
drop_cls(phylota, cid)
}
\arguments{
\item{phylota}{Phylota object}

\item{cid}{Cluster ID(s) to be kept}
}
\value{
phylota
}
\description{
Drops all clusters except those
identified by user.
}
