% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-tools-phylota-drop.R
\name{drop_sqs}
\alias{drop_sqs}
\title{Drop sequences in a cluster}
\usage{
drop_sqs(phylota, cid, sid)
}
\arguments{
\item{phylota}{Phylota object}

\item{cid}{Cluster ID}

\item{sid}{Sequence ID(s) to be kept}
}
\value{
phylota
}
\description{
Drop all sequences in a cluster
except those identified by user.
}
