% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clstrsobj-tools.R
\name{drpClstrs}
\alias{drpClstrs}
\title{Drop clusters from clusters object}
\usage{
drpClstrs(clstrs_obj, id)
}
\arguments{
\item{clstrs_obj}{Clusters Object}

\item{id}{Clusters IDs to be kept}
}
\description{
Drop clusters from clusters object. Only
the ids provided are kept.
}
