% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-tools-phylota.R
\name{drpSqs}
\alias{drpSqs}
\title{Drop sequences from cluster}
\usage{
drpSqs(clstrs_obj, cid, sid)
}
\arguments{
\item{clstrs_obj}{Clusters Object}

\item{cid}{Clusters ID}

\item{sid}{Sequence IDs to be kept}
}
\description{
Drop sequences from a cluster
}
