% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-tools.R
\name{dwnld}
\alias{dwnld}
\title{Download sequences for txids}
\usage{
dwnld(wd, txids, phylt_nds, mdl_thrshld, mx_sq_lngth, verbose)
}
\arguments{
\item{txids}{Taxonomic node IDs, numeric vector}

\item{phylt_nds}{PhyLoTa data.frame}

\item{mdl_thrshld}{Maximum number of sequences per txid}

\item{mx_sq_lngth}{Maximum sequence length}

\item{verbose}{Verbose? T/F}
}
\description{
Look up and download all sequences for
given taxonomic IDs.
}
\details{
Sequence downloads are cached.
}
