% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi-tools.R
\name{dwnldFrmNCBI}
\alias{dwnldFrmNCBI}
\title{dwnldFrmNCBI}
\usage{
dwnldFrmNCBI(txid, direct = FALSE, mx_lngth = 25000, mx_sqs = 1e+05,
  verbose = FALSE)
}
\arguments{
\item{txid}{NCBI taxon identifier}

\item{direct}{Whether to download all 'directly' linked
sequences ([Organism:noexp] in Genbank search) or all 'subtree'
sequences ([Organism:exp])}

\item{mx_lngth}{Maximum length of sequence to be included in GI list}

\item{mx_sqs}{Take a random subset of mx_sqs if the number of
seqs exceeds it.
Only makes sense when direct=TRUE}
}
\value{
list of lists containing sequence objects
}
\description{
Given a taxon ID, queries NCBI
for sequences that match the taxid.
}
