% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage-taxise.R
\name{dwnldTxRcrds}
\alias{dwnldTxRcrds}
\title{Download and parse taxonomic records}
\usage{
dwnldTxRcrds(txids, ps)
}
\arguments{
\item{txids}{Vector of taxonomic IDs}

\item{ps}{Parameter list}
}
\value{
List of taxonomic records
}
\description{
Downloads taxonomic records from a vector
of IDs. Parses the returned records to control for
lowest rank.
}
