% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-tools.R
\name{fltr}
\alias{fltr}
\title{Get all node IDs that will be processed}
\usage{
fltr(txid, phylt_nds, mdl_thrshld, mx_blst_sqs, verbose = FALSE)
}
\description{
All nodes for which all children
contain < mx_blst_sqs sequences
}
