% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast-tools.R
\name{fltrBlstRs}
\alias{fltrBlstRs}
\title{Filter BLAST results}
\usage{
fltrBlstRs(blst_rs, min_cvrg = 0.51, verbose = FALSE)
}
\arguments{
\item{blst_rs}{BLAST results}
}
\description{
Can you fill this in Hannes?
}
