% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylota-tools.R
\name{fltrClstrSqs}
\alias{fltrClstrSqs}
\title{Filter out sequences for a cluster}
\usage{
fltrClstrSqs(clstrs_obj, id, rank = "species", mn_pambg = 0.1)
}
\arguments{
\item{clstrs_obj}{Clusters Object}

\item{id}{Clusters ID}

\item{rank}{Taxonomic rank}

\item{mn_pambg}{Min. ambiguous bases}
}
\description{
Identify all sequences in a cluster
 keep only those sequences representing rank of choice (e.g. species)
 identify all sequences of the same taxa
 select 'best' sequence for each taxon
  - drop all with too much ambiguity
  - select longest sequence per taxon
}
