% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-tools-clrcrdbx.R
\name{genClRcrdBx}
\alias{genClRcrdBx}
\title{Generate Box of Cluster Records}
\usage{
genClRcrdBx(clstr_rcrds)
}
\arguments{
\item{clstr_rcrds}{Cluster records}
}
\description{
Takes a list of ClRcrds, returns a box.
}
