% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage3-tools.R
\name{genClRcrds}
\alias{genClRcrds}
\alias{clusterrec_gen}
\title{Generate list of clusters}
\usage{
clusterrec_gen(cluster_list, txid, sqs, typ)
}
\arguments{
\item{txid}{Taxonomic node ID}

\item{sqs}{Sequnece records}

\item{typ}{Subtree of direct?}

\item{cluster_lst}{List of list of cluster descriptions}
}
\description{
Takes a list of lists of cluster descriptions,
returns a ClRcrdBx.
}
