% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage-tools-cluster.R
\name{genClstr}
\alias{genClstr}
\title{Generate list of Clstrs}
\usage{
genClstr(clstr_lst, txid, sqs, typ)
}
\arguments{
\item{clstr_lst}{List of list of cluster descriptions}

\item{txid}{Taxonomic node ID}

\item{sqs}{Sequnece records}

\item{typ}{Subtree of direct?}
}
\description{
Takes a list of lists of cluster descriptions,
returns a list of Clstr objects.
}
