% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage-taxise.R
\name{genPhylotaNds}
\alias{genPhylotaNds}
\title{Generate taxonomic nodes in PhyLoTa format}
\usage{
genPhylotaNds(nid_sets, td_nds, td_nms, ps)
}
\arguments{
\item{nid_sets}{Taxonomic processable and unprocessable
node IDs, list as generated by \code{getMngblIds}}

\item{v}{v? T/F}
}
