% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-sqs.R
\name{genSqRcrd}
\alias{genSqRcrd}
\title{Generate SqRcrd}
\usage{
genSqRcrd(accssn, gi, nm, txid, sq, dfln, orgnsm, ml_typ, rcrd_typ, vrsn, age,
  lctn = NULL)
}
\arguments{
\item{accssn}{Accession ID}

\item{gi}{GI}

\item{nm}{Sequence name}

\item{txid}{Taxonomic ID of source organism}

\item{sq}{Sequence}

\item{dfln}{Defintion line}

\item{orgnsm}{Source organism name}

\item{ml_typ}{Molecule type}

\item{rcrd_typ}{Sequence record type}

\item{vrsn}{Accession version}

\item{age}{Number of days since upload}

\item{lctn}{Location numbers for features, e.g. '1..200'}
}
\description{
Creates an S4 SqRcrd
}
