% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxise-tools.R
\name{genTDObj}
\alias{genTDObj}
\title{Generate R object from NCBI taxonomy dump}
\usage{
genTDObj()
}
\description{
Generates an interrogable R object
from the NCBI taxonomy dump using the \code{CHNOSZ}
library. The returned object if a list of taxonomic nodes
and names.
}
