% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage-taxise.R
\name{genTxDct}
\alias{genTxDct}
\title{Generate taxonomic dictionary}
\usage{
genTxDct(txids, rcrds)
}
\arguments{
\item{txids}{Vector of taxonomic IDs}

\item{rcrds}{List of taxonomic records}
}
\value{
TxDct
}
\description{
Takes a vector of txids and a list
of taxonomic records and returns a taxonomic dictionary.
}
