% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxise.R
\name{genTxNds}
\alias{genTxNds}
\title{Extract and filter NCBI taxonomy}
\usage{
genTxNds(wd)
}
\arguments{
\item{wd}{Working directory}
}
\description{
Look up NCBI taxonomy and generate
a list of descendent taxonomic nodes based on
provided root \code{txid}. Descendent nodes
are written to disk in PhyLoTa table format.
This is a necessary step before clustering can
be performed.
}
\details{
Object will be cached.
}
