% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{get.manageable.node.set}
\alias{get.manageable.node.set}
\title{Get manageable node set}
\usage{
get.manageable.node.set(root.taxa, nodes, max.descendants = 10000,
  timeout = 10, nodesfile = "")
}
\description{
TODO
}
\examples{
# TODO
}
