% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-tools-phylota-get.R
\name{get_cl_slot}
\alias{get_cl_slot}
\title{Get slot data for each cluster}
\usage{
get_cl_slot(phylota, cid, slt_nm = list_clrcrd_slots())
}
\arguments{
\item{phylota}{Phylota object}

\item{cid}{Cluster ID}

\item{slt_nm}{Slot name}
}
\value{
vector
}
\description{
Get slot data for cluster(s)
}
