% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-get.R
\name{get_nsqs}
\alias{get_nsqs}
\title{Count number of sequences}
\usage{
get_nsqs(phylota, cid)
}
\arguments{
\item{phylota}{Phylota object}

\item{cid}{CLuster ID(s)}
}
\value{
vector
}
\description{
Count the number of sequences in a cluster(s)
}
