% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-get.R
\name{get_ntaxa}
\alias{get_ntaxa}
\title{Count number of unique taxa}
\usage{
get_ntaxa(phylota, cid = NULL, sid = NULL, rnk = NULL,
  keep_higher = FALSE)
}
\arguments{
\item{phylota}{Phylota object}

\item{cid}{CLuster ID(s)}

\item{sid}{Sequence ID(s)}

\item{rnk}{Taxonomic rank}

\item{keep_higher}{Keep higher taxonomic ranks?}
}
\value{
vector
}
\description{
Count the number of unique taxa
represented by cluster(s) or sequences in phylota table
Use rnk to specify a taxonomic level to count. If NULL
counts will be made to the lowest level reported on NCBI.
}
