% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-get.R
\name{get_sq_slot}
\alias{get_sq_slot}
\title{Get slot data for each sequence}
\usage{
get_sq_slot(phylota, cid = NULL, sid = NULL, slt_nm = list_sqrcrd_slots())
}
\arguments{
\item{phylota}{Phylota object}

\item{cid}{Cluster ID}

\item{sid}{Sequence ID(s)}

\item{slt_nm}{Slot name}
}
\value{
vector
}
\description{
Get slot data for either or sequences
in a cluster of a vector of sequence IDs.
Use list_sqrcrd_slots() for a list of available
slots.
}
