% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-get.R
\name{get_tx_slot}
\alias{get_tx_slot}
\title{Get slot data for each taxon record}
\usage{
get_tx_slot(phylota, txid, slt_nm = list_taxrec_slots())
}
\arguments{
\item{phylota}{Phylota object}

\item{txid}{Taxonnomic ID}

\item{slt_nm}{Slot name}
}
\value{
vector or list
}
\description{
Get slot data for taxa(s)
}
