% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-tools-phylota-get.R
\name{get_txids}
\alias{get_txids}
\title{Get taxonomic IDs}
\usage{
get_txids(phylota, cid = NULL, sid = NULL, rnk = NULL,
  keep_higher = FALSE)
}
\arguments{
\item{phylota}{Phylota object}

\item{cid}{CLuster ID}

\item{sid}{Sequence ID(s)}

\item{rnk}{Taxonomic rnk}

\item{keep_higher}{Keep higher taxonomic IDs?}
}
\value{
vector
}
\description{
Return taxonomic IDs for
a vector of sequence IDs or all sequences in a cluster.
User can specify what rank the IDs should be
returned. If NULL, the lowest level is returned.
}
\details{
If keep_higher is TRUE, any sequence that has
a identity that is higher than the given rank will be returned.
If FALSE, these sequences will return ''.
}
