% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-tools-txdct.R
\name{getADs}
\alias{getADs}
\title{Get all descendants}
\usage{
getADs(id, txdct)
}
\arguments{
\item{id}{txid}

\item{txdct}{TxDct}
}
\value{
Vector
}
\description{
Look-up all taxonomic descendants of a node from TxDct.
}
