% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylota-tools.R
\name{getBestClstrs}
\alias{getBestClstrs}
\title{Get best clusters}
\usage{
getBestClstrs(clstrs_obj, n)
}
\arguments{
\item{clstrs_obj}{Clusters Object}

\item{n}{Number of cluster IDs to return}
}
\description{
TODO
}
