% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clstrsobj-tools.R
\name{getClGCR}
\alias{getClGCR}
\title{Return mean GC-ratio of sequences in cluster}
\usage{
getClGCR(clstrs_obj, id)
}
\arguments{
\item{clstrs_obj}{clstrs_obj}

\item{id}{cluster ID(s)}
}
\description{
Return the mean proportion of G or C in
the unambiguous sequence of all sequences in cluster(s)
}
