% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-tools.R
\name{getDDFrmPhyltNds}
\alias{getDDFrmPhyltNds}
\title{Get direct descendants from PhyLoTa nodes}
\usage{
getDDFrmPhyltNds(txid, phylt_nds)
}
\arguments{
\item{txid}{parent}

\item{phylt_nds}{PhyLoTa nodes data.frame}
}
\description{
Find next node IDs using the PhyLoTa
nodes data.frame.
}
\details{
Returns vector of node IDs
}
