% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-tools-txdct.R
\name{getDDs}
\alias{getDDs}
\title{Get direct descendants}
\usage{
getDDs(id, txdct)
}
\arguments{
\item{id}{txid}

\item{txdct}{TxDct}
}
\value{
Vector
}
\description{
Look-up direct taxonomic descendants of a node from TxDct.
}
