% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unused.R
\name{getGnsFrmPhyltNds}
\alias{getGnsFrmPhyltNds}
\title{Get genus for txid using PhyLoTa nodes}
\usage{
getGnsFrmPhyltNds(txid, phylt_nds)
}
\arguments{
\item{txid}{Taxonomic node ID, numeric}

\item{phylt_nds}{PhyLoTa nodes data.frame}
}
\description{
Returns genus for a taxonomic node ID
}
