% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-tools-txdct.R
\name{getIDFrmTxdct}
\alias{getIDFrmTxdct}
\title{Get taxonomic IDs by rank}
\usage{
getIDFrmTxdct(txdct, id, ret = c("TaxId", "ScientificName"),
  rank = c("superkingdom", "kingdom", "phylum", "subphylum", "class",
  "superorder", "order", "suborder", "infraorder", "parvorder", "family",
  "genus", "species", "subspecies"))
}
\arguments{
\item{txdct}{Taxonomic dictionary}

\item{id}{Taxon IDs}

\item{ret}{Return ID or Name?}

\item{rank}{Rank of output}
}
