% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxise-tools.R
\name{getMngblIds}
\alias{getMngblIds}
\title{Identify manageable taxonomic node IDs}
\usage{
getMngblIds(txid, td_nds, ps)
}
\arguments{
\item{txid}{Taxomomic IDs}

\item{td_nds}{'nds' element from the \code{tdobj}}

\item{ps}{Parameters}
}
\description{
Given a root \code{txid}, return a set of
taxonomic node IDs that only have up to a maximum number
of descendants.
}
\details{
Returns a \code{list}.
If there are more descendants than mx_dscndnts or
retrieveing the number takes longer than \code{tmout}
seconds, the taxonomic ID is discarded and its children
are added to the queue.
}
