% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.R
\name{getNdKids}
\alias{getNdKids}
\title{Get children IDs}
\usage{
getNdKids(tree, id)
}
\arguments{
\item{tree}{\code{TreeMan} object}

\item{id}{node id}
}
\description{
Return the node ids of all tips that descend from node.
}
\details{
Returns a vector
}
\examples{

tree <- randTree(10)
# everyone descends from root
getNdKids(tree, id = tree["root"])
}
\seealso{
\code{\link{getNdsKids}},
\url{https://github.com/DomBennett/treeman/wiki/get-methods}
}
