% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.R
\name{getNdLng}
\alias{getNdLng}
\title{Get lineage}
\usage{
getNdLng(tree, id)
}
\arguments{
\item{tree}{\code{TreeMan} object}

\item{id}{node id}
}
\description{
Return unique taxonomic names for connecting \code{id} to root.
}
\details{
Returns a vector.
}
\examples{

data(mammals)
# return human lineage
getNdLng(mammals, id = "Homo_sapiens")
}
\seealso{
\code{\link{getNdsLng}}, \code{\link{getNdsFrmTxnyms}},
\url{https://github.com/DomBennett/treeman/wiki/get-methods}
}
