% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.R
\name{getNdPrdst}
\alias{getNdPrdst}
\title{Get pre-distance}
\usage{
getNdPrdst(tree, id)
}
\arguments{
\item{tree}{\code{TreeMan} object}

\item{id}{node id}
}
\description{
Return root to tip distance (prdst) for \code{id}
}
\details{
Sums the lengths of all branches from \code{id} to root.
}
\examples{

tree <- randTree(10)
getNdPrdst(tree, id = "t1") # return the distance to root from t1
}
\seealso{
\code{\link{getNdsPrdst}},
\url{https://github.com/DomBennett/treeman/wiki/get-methods}
}
