% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.R
\name{getNdsFrmTxnyms}
\alias{getNdsFrmTxnyms}
\title{Get IDs for nodes represented txnyms}
\usage{
getNdsFrmTxnyms(tree, txnyms)
}
\arguments{
\item{tree}{\code{TreeMan} object}

\item{txnyms}{vector of taxonomic group names}
}
\description{
Return a list of IDs for any node that contains the given txnyms.
}
\details{
Returns a list. Txnyms must be spelt correctly.
}
\examples{

data(mammals)
# what ID represents the apes?
getNdsFrmTxnyms(mammals, "Hominoidea")
}
\seealso{
\code{\link{taxaResolve}}, \code{\link{setTxnyms}}, \code{\link{searchTxnyms}},
\code{\link{getNdsLng}}, \code{\link{getNdLng}}
}
