% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.R
\name{getNdSstr}
\alias{getNdSstr}
\title{Get sister id}
\usage{
getNdSstr(tree, id)
}
\arguments{
\item{tree}{\code{TreeMan} object}

\item{id}{node id}
}
\description{
Returns the id of the sister(s) of node id given.
}
\details{
An error is raised if there is no sister (e.g. for the root).
 There can be more than one sister if tree is polytomous.
}
\examples{

tree <- randTree(10)
getNdSstr(tree, id = "t1")
}
\seealso{
\code{\link{getNdsSstr}},
\url{https://github.com/DomBennett/treeman/wiki/get-methods}
}
