% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster2-tools.R
\name{getSeedSqs}
\alias{getSeedSqs}
\title{Extract seed sequences from clusters}
\usage{
getSeedSqs(clstrs, sqs)
}
\arguments{
\item{clstrs}{All clusters}

\item{sqs}{All seed sequences to be BLASTed}
}
\description{
Returns seed sequences of all clusters
}
