% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-tools-txdct.R
\name{getSngltns}
\alias{getSngltns}
\title{Get equivalent singleton IDs}
\usage{
getSngltns(txid, txdct)
}
\arguments{
\item{txid}{txid}

\item{txdct}{TxDct}
}
\value{
Character vector
}
\description{
For a given taxonomic ID, return all the
equivalent singleton IDs.
}
\details{
phylotaR drops all singleton nodes from
the NCBI taxonomy upon initiation. To retrieve all
the singleton IDs for an ID use this function.
Singletons are defined as nodes in the hierarchy
that do not split.
}
