% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.R
\name{getSpnAge}
\alias{getSpnAge}
\title{Get age range}
\usage{
getSpnAge(tree, id, tree_age)
}
\arguments{
\item{tree}{\code{TreeMan} object}

\item{id}{node id}

\item{tree_age}{numeric value of known age of tree}
}
\description{
Return start and end ages for \code{id} from when it first appears to when it splits
}
\details{
Returns a dataframe.
}
\examples{

data(mammals)
# mammal_age <- getAge(mammals)  # ~166.2, needs to be performed when tree is not up-to-date
getSpnAge(mammals, id = "Homo_sapiens", tree_age = 166.2)
}
\seealso{
\code{\link{getNdAge}},
\code{\link{getNdsAge}},
\code{\link{getSpnsAge}},
\url{https://github.com/DomBennett/treeman/wiki/get-methods}
}
