% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-co-sq.R
\name{getSqAmbs}
\alias{getSqAmbs}
\title{Return sequence deflines}
\usage{
getSqAmbs(clstrs_obj, id)
}
\arguments{
\item{clstrs_obj}{clstrs_obj}

\item{id}{cluster ID(s)}
}
\description{
Return all deflines for a cluster's sequences
}
