% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co-sq-tools.R
\name{getSqGCR}
\alias{getSqGCR}
\title{Return GC-ratio by sequences}
\usage{
getSqGCR(clstrs_obj, id)
}
\arguments{
\item{clstrs_obj}{clstrs_obj}

\item{id}{cluster ID(s)}
}
\description{
Return the proportion of G or C in
the unambiguous sequence of all sequences in cluster
}
