% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-co-sq.R
\name{getSqTx}
\alias{getSqTx}
\title{Return taxonomic ID per sequence}
\usage{
getSqTx(clstrs_obj, cid = NULL, sid = NULL, rank = FALSE)
}
\arguments{
\item{clstrs_obj}{clstrs_obj}

\item{id}{cluster ID(s)}
}
\description{
Get the taxonomic ID by sequence in
a cluster.
}
