% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxise-tools.R
\name{getStats}
\alias{getStats}
\title{Generate PhyLoTa statistics for txid}
\usage{
getStats(txid, phylt_nds, td_nds, td_nms, ps, rcrsv = FALSE)
}
\arguments{
\item{txid}{Taxonomic ID}

\item{phylt_nds}{PhyLoTa data.frame}

\item{td_nds}{'nds' element from the \code{tdobj}}

\item{td_nms}{'nms' element from the \code{tdobj}}

\item{ps}{Parameters}

\item{rcrsv}{Add stats for kid nodes recursively? T/F}
}
\details{
Returns an updated \code{phylt_nds}.
}
