% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clstrsobj-tools.R
\name{getTxData}
\alias{getTxData}
\title{Return taxonomic information}
\usage{
getTxData(clstrs_obj, txid, sltnm = "scientificname")
}
\arguments{
\item{clstrs_obj}{clstrs_obj}

\item{txid}{Taxonomic ID(s)}

\item{sltnm}{Taxonomic data slot name}
}
\description{
Return vector of taxonomic data values
using taxonomic IDs from a clsuters object.
}
