% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage-taxise.R
\name{getTxids}
\alias{getTxids}
\title{Searches for descendent taxonomic IDs}
\usage{
getTxids(ps, retmax = 10000)
}
\arguments{
\item{ps}{Parameter list}

\item{retmax}{integer, maximum number of IDs to return per query}
}
\value{
Vector of txids
}
\description{
Searches NCBI taxonomy for all descendent
taxonomic nodes.
}
