% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.R
\name{getUnqNds}
\alias{getUnqNds}
\title{Get unique nodes represented by tips}
\usage{
getUnqNds(tree, tids)
}
\arguments{
\item{tree}{\code{TreeMan} object}

\item{tids}{vector of tip IDs}
}
\description{
Return a list of IDs for any node that are represented by tip IDs given.
}
\details{
Returns a vector.
}
\examples{

tree <- randTree(10)
unqnds <- getUnqNds(tree, c("t1", "t2"))
}
\seealso{
\code{\link{getCnnctdNds}}, \code{\link{calcFrPrp}},
\code{\link{calcPhyDv}}
}
