% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage-tools-download.R
\name{hrrchcDwnld}
\alias{hrrchcDwnld}
\title{Hierarchically get sequences for a txid}
\usage{
hrrchcDwnld(txid, txdct, ps, lvl = 0)
}
\arguments{
\item{txid}{Taxonomic node ID, numeric}
}
\description{
Looks up and downloads sequences for a
taxonomic ID.
}
