% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache-tools.R
\name{intPrgrss}
\alias{intPrgrss}
\title{Initialise progress list in cache}
\usage{
intPrgrss(wd)
}
\arguments{
\item{wd}{Working directory}
}
\description{
Creates a progress list recording each
stage run in cache
}
