% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-tools-phylota-special.R
\name{is_txid_in_cl}
\alias{is_txid_in_cl}
\title{Is txid in cluster?}
\usage{
is_txid_in_cl(phylota, txid, cid)
}
\arguments{
\item{phylota}{PhyLoTa object}

\item{txid}{Taxonomic ID}

\item{cid}{Cluster ID}
}
\value{
boolean
}
\description{
Checks if given txid is represented
by any of the sequences of a cluster by searching
through all the sequence search organism lineages.
}
\examples{
data(tinamous)
cid <- tinamous@cids[[1]]
cl <- tinamous[[cid]]
sq <- tinamous[[cl@sids[[1]]]]
txid <- sq@txid
# expect true
is_txid_in_cl(phylota=tinamous, txid=txid, cid=cid)
}
