% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-tools-phylota-special.R
\name{is_txid_in_sq}
\alias{is_txid_in_sq}
\title{Is txid in sequence?}
\usage{
is_txid_in_sq(phylota, txid, sid)
}
\arguments{
\item{phylota}{PhyLoTa object}

\item{txid}{Taxonomic ID}

\item{sid}{Sequence ID}
}
\value{
boolean
}
\description{
Checks if given txid is represented
by sequence by looking at sequence source organism's
lineage.
}
\examples{
data(tinamous)
sid <- tinamous@sids[[1]]
sq <- tinamous[[sid]]
txid <- sq@txid
# expect true
is_txid_in_sq(phylota=tinamous, txid=txid, sid=sid)
}
