% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.R
\name{isUltrmtrc}
\alias{isUltrmtrc}
\title{Is tree ultrametric?}
\usage{
isUltrmtrc(tree, tol = 1e-08)
}
\arguments{
\item{tree}{\code{TreeMan} object}

\item{tol}{zero tolerance}
}
\description{
Return TRUE if all tips end at 0, else FALSE.
}
\details{
Returns a boolean. This function works in the background
for the \code{['ultr']} slot in a \code{TreeMan} object.
}
\examples{

tree <- randTree(10)
(isUltrmtrc(tree))
}
\seealso{
\code{\link{getLvng}}, \code{\link{getDcsd}}
}
