% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-tools-clrcrdbx.R
\name{jnBxs}
\alias{jnBxs}
\title{Join two cluster record boxes}
\usage{
jnBxs(bx_1, bx_2)
}
\arguments{
\item{bx_1}{Cluster record box}

\item{bx_2}{Cluster record box}
}
\description{
Takes two ClRcrdBxs and joins them into
a single ClRcrdBx
}
