% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster2-tools.R
\name{jnClstrs}
\alias{jnClstrs}
\title{Join clusters for merging}
\usage{
jnClstrs(blst_rs, seed_ids, all_clstrs)
}
\arguments{
\item{blst_rs}{Seed sequence BLAST results}

\item{seed_ids}{Seed sequence IDs}

\item{all_clstrs}{List of all clusters}
}
\description{
Uses seed sequence BLAST results and IDs
to join clusters identifed as sisters into single clusters.
Resulting object is of joined clusters, merging is required
to reformat the clusters for subsequent analysis.
}
