% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage4-tools.R
\name{jnclusters}
\alias{jnclusters}
\alias{clusters_join}
\title{Join clusters for merging}
\usage{
clusters_join(blast_res, seed_ids, all_clusters, ps)
}
\arguments{
\item{seed_ids}{Seed sequence IDs}

\item{all_clusters}{List of all clusters}

\item{ps}{Parameters}

\item{blst_rs}{Seed sequence BLAST results}
}
\description{
Uses seed sequence BLAST results and IDs
to join clusters identifed as sisters into single clusters.
Resulting object is of joined clusters, merging is required
to reformat the clusters for subsequent analysis.
}
