% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-cache.R
\name{ldGIs}
\alias{ldGIs}
\alias{sids_load}
\title{Load GIs from cache}
\usage{
sids_load(wd, txid)
}
\arguments{
\item{wd}{Working directory}

\item{txid}{Taxonomic ID, numeric}
}
\description{
Load GIs downloaded by
\code{getGIs} function.
}
