% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-cache.R
\name{ldNcbiCch}
\alias{ldNcbiCch}
\alias{ncbicache_load}
\title{Retrieve cached NCBI query}
\usage{
ncbicache_load(fnm, args, wd)
}
\arguments{
\item{fnm}{NCBI Entrez function name}

\item{args}{Args used for function}

\item{wd}{Working directory}
}
\description{
Run this function to load cached NCBI queries.
}
