% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache-tools.R
\name{ldObj}
\alias{ldObj}
\title{Load a named object from the cache}
\usage{
ldObj(wd, nm)
}
\arguments{
\item{wd}{Working directory}

\item{nm}{Object name}
}
\description{
Loads an object from the cache
as stored by \code{svObj}
}
