% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-cache.R
\name{ldPrmtrs}
\alias{ldPrmtrs}
\title{Load paramters from cache}
\usage{
ldPrmtrs(wd)
}
\arguments{
\item{wd}{Working directory}
}
\description{
Parameters are held in cache, use
this function to load parameters set for a wd.
}
