% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylotaR.R
\name{list_clrcrd_slots}
\alias{list_clrcrd_slots}
\alias{list_clusterrec_slots}
\title{List all ClRcrd slots}
\usage{
list_clusterrec_slots()
}
\value{
vector
}
\description{
Returns a vector of all
available ClRcrd slots.
}
