% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylotaR.R
\name{list_sqrcrd_slots}
\alias{list_sqrcrd_slots}
\alias{list_seqrec_slots}
\title{List all SqRcrd slots}
\usage{
list_seqrec_slots()
}
\value{
vector
}
\description{
Returns a vector of all
available SqRcrd slots.
}
