% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylotaR.R
\name{list_txrcrd_slots}
\alias{list_txrcrd_slots}
\alias{list_taxrec_slots}
\title{List all TxRcrd slots}
\usage{
list_taxrec_slots()
}
\value{
vector
}
\description{
Returns a vector of all
available TxRcrd slots.
}
