% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-blast.R
\name{mkBlstDB}
\alias{mkBlstDB}
\alias{blastdb_gen}
\title{Make a BLAST database}
\usage{
blastdb_gen(sqs, dbfl, ps)
}
\arguments{
\item{sqs}{Sequences}

\item{dbfl}{Outfile for database}

\item{ps}{Parameters}
}
\description{
Generate BLAST database in wd 
for given sequences.
}
