% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-entrez.R
\name{mkSrchTrm}
\alias{mkSrchTrm}
\alias{searchterm_gen}
\title{Construct GenBank Search Term}
\usage{
searchterm_gen(txid, ps, drct = FALSE)
}
\arguments{
\item{txid}{Taxonomic ID}

\item{ps}{Parameter list}

\item{drct}{Node-level only or subtree as well? Default FALSE.}
}
\description{
Construct search term for searching
GenBank's nucleotide database.
Limits the maximum size of sequences, avoids
whole genome shotguns, predicted, unverified and
synthetic sequences.
}
