% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage4-tools.R
\name{mrgclusters}
\alias{mrgclusters}
\alias{clusters_merge}
\title{Merge joined clusters}
\usage{
clusters_merge(jnd_clusters, txdct)
}
\arguments{
\item{jnd_clusters}{List of joined cluster records}

\item{txdct}{Taxonomic dictionary}
}
\description{
Takes a list of joined clusters and computes
each data slot to create a single merged cluster.
txdct is required for parent look-up.
}
