% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.R
\name{multiPhylo-to-TreeMen}
\alias{multiPhylo-to-TreeMen}
\title{Convert multiPhylo to TreeMen}
\description{
Return a \code{TreeMen} from ape's \code{mutlPhylo}
}
\examples{

library(ape)
trees <- c(rtree(10), rtree(10), rtree(10))
trees <- as(trees, "TreeMen")
}
\seealso{
\code{\link{TreeMan-to-phylo}},
\code{\link{phylo-to-TreeMan}},
\code{\link{TreeMen-to-multiPhylo}}
\code{\link{TreeMan-class}}
}
