% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R, R/taxise-tools.R
\name{nDscndnts}
\alias{nDscndnts}
\alias{nDscndnts}
\title{Count descendants}
\usage{
nDscndnts(id, td_nds)

nDscndnts(id, td_nds)
}
\arguments{
\item{id}{Taxonomic ID}

\item{td_nds}{NCBI taxonomic nodes}
}
\description{
TODO

Count the number of children
descending from a node in the NCBI taxonomy
dump.
}
\examples{
# TODO
}
