% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi-tools.R
\name{nNcbiNds}
\alias{nNcbiNds}
\title{Count number of descending taxonomic nodes}
\usage{
nNcbiNds(txid, ps)
}
\arguments{
\item{txid}{Taxonomic ID}

\item{ps}{Parameters}
}
\description{
Searches NCBI taxonomy and returns
number of descendents taxonomic nodes (species, genera ...)
of ID.
}
