% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi-tools.R
\name{nSqs}
\alias{nSqs}
\title{Count number of sequences for txid}
\usage{
nSqs(txid, ps, direct = FALSE)
}
\arguments{
\item{txid}{Taxonomic ID}

\item{direct}{exp or noexp? T/F}
}
\description{
Return the number of sequences
associated with a taxonomic ID on NCBI GenBank.
}
