% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{parameters_reset}
\alias{parameters_reset}
\title{Change parameters in a working directory}
\usage{
parameters_reset(wd, parameters, values)
}
\arguments{
\item{wd}{Working directory}

\item{parameters}{Parameters to be changed}

\item{values}{New values for each parameter}
}
\description{
Reset parameters afater running \code{setUp()}.
}
