% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline-tools.R
\name{parameters_setup}
\alias{parameters_setup}
\title{Set Up Parameters}
\usage{
parameters_setup(wd, ncbi_execs, ...)
}
\arguments{
\item{wd}{Working directory}

\item{ncbi_execs}{File directories for NCBI tools, see \code{blast_setup()}}

\item{...}{Set parameters, see parameters()}
}
\description{
Initiates cache of parameters.
}
\seealso{
\link{parameters}, \link{blast_setup}
}
