% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.R
\name{phylo-to-TreeMan}
\alias{phylo-to-TreeMan}
\title{Convert phylo to TreeMan}
\description{
Return a \code{TreeMan} from ape's \code{phylo}. This will
preserve node labels, if they are a alphanumeric.
}
\examples{

library(ape)
tree <- compute.brlen(rtree(10))
tree <- as(tree, "TreeMan")
}
\seealso{
\code{\link{TreeMan-to-phylo}},
\code{\link{TreeMen-to-multiPhylo}}
\code{\link{multiPhylo-to-TreeMen}}
\code{\link{TreeMan-class}}
}
