% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-tools-phylota-special.R
\name{plot_phylota_pa}
\alias{plot_phylota_pa}
\title{Plot presence/absence matrix}
\usage{
plot_phylota_pa(phylota, cids, txids, cnms = cids, txnms = txids)
}
\arguments{
\item{phylota}{PhyLoTa object}

\item{cids}{Vector of cluster IDs}

\item{txids}{Vector of taxonomic IDs}

\item{cnms}{Cluster names}

\item{txnms}{Taxonomic names}
}
\value{
geom_object
}
\description{
Plot presence/absence of taxa by each
cluster in phylota object.
}
\details{
Cluster names and taxonomic names
can be given to the function, by default IDs are used.
}
\examples{
library(phylotaR)
data(cycads)
# drop all but first ten
cycads <- drop_cls(cycads, cycads@cids[1:10])
# plot all
p <- plot_phylota_pa(phylota=cycads, cids=cycads@cids,
                     txids=cycads@txids)
print(p)  # lots of information, difficult to interpret
# get genus-level taxonomic names
genus_txids <- get_txids(cycads,
                         txids=cycads@txids,
                         rnk='genus')
genus_txids <- unique(genus_txids)
# dropping missing
genus_txids <- genus_txids[genus_txids != '']
genus_nms <- get_tx_slot(cycads, genus_txids, slt_nm='scnm')
# make alphabetical for plotting
genus_nms <- sort(genus_nms, decreasing=TRUE)
# generate geom_object
p <- plot_phylota_pa(phylota=cycads, cids=cycads@cids,
                     txids=genus_txids, txnms=genus_nms)
# plot
print(p)  # easier to interpret
}
