% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clstrsobj-tools.R
\name{plotTable}
\alias{plotTable}
\title{Plot table of txid presence by clusters}
\usage{
plotTable(clstrs_obj, clstr_ids)
}
\arguments{
\item{clstrs_obj}{clstrs_obj}

\item{clstr_ids}{IDs of clusters in table}
}
\description{
TODO
}
