% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-entrez.R
\name{prtDwnld}
\alias{prtDwnld}
\title{prtDwnld}
\usage{
prtDwnld(gis, ps)
}
\arguments{
\item{gis}{Sequence GI IDs}

\item{ps}{Parameter list}
}
\value{
Vector of sequence objects
}
\description{
Download batch of sequences.
}
\details{
Given a set of IDs, downloads and breaks
up by feature information.
}
