% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage-tools-taxise.R
\name{prtDwnldTxRcrds}
\alias{prtDwnldTxRcrds}
\title{Download Taxonomic Records}
\usage{
prtDwnldTxRcrds(txids, ps)
}
\arguments{
\item{txids}{Vector of taxonomic IDs}

\item{ps}{Parameter list}
}
\value{
list of list
}
\description{
Downloads one batch of taxonomic
records.
}
