% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.R
\name{pstMnp}
\alias{pstMnp}
\title{Update prinds and tinds}
\usage{
pstMnp(tree)
}
\arguments{
\item{tree}{\code{TreeMan} object}
}
\description{
Return tree with updated slots.
}
\details{
This function is automatically run. Only run, if you
are creating yor own functions to add and remove elements of the
\code{ndlst}.
}
\seealso{
\code{\link{updateSlts}}, \code{\link{addNdmtrx}},
\code{\link{getAge}}
}
