% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-cache.R
\name{rdPrgrss}
\alias{rdPrgrss}
\alias{progress_read}
\title{Read the progress from cache}
\usage{
progress_read(wd)
}
\arguments{
\item{wd}{Working directory}
}
\description{
Return the last completed stage using the cache.
}
