% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-special.R
\name{read_phylota}
\alias{read_phylota}
\title{Generate a PhyLoTa object in R}
\usage{
read_phylota(wd)
}
\arguments{
\item{wd}{Working directory}
}
\value{
phylota
}
\description{
Creates a PhyLoTa object containing
information on clusters, sequences and taxonomy
from the working directory of a completed pipeline.
}
