% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-tools-phylota.R
\name{readPhyLoTa}
\alias{readPhyLoTa}
\title{Generate a PhyLoTa object in R}
\usage{
readPhyLoTa(wd)
}
\arguments{
\item{wd}{Working directory}
}
\description{
Creates a PhyLoTa object containing
information on clusters, sequences and taxonomy
from the working directory of a completed pipeline.
}
