% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes.R
\docType{class}
\name{RecordCluster-class}
\alias{RecordCluster-class}
\alias{RecordCluster-method}
\alias{as.character,RecordCluster-method}
\alias{show,RecordCluster-method}
\alias{print,RecordCluster-method}
\alias{str,RecordCluster-method}
\alias{summary,RecordCluster-method}
\title{RecordCluster-class}
\usage{
\S4method{as.character}{RecordCluster}(x)

\S4method{show}{RecordCluster}(object)

\S4method{print}{RecordCluster}(x)

\S4method{str}{RecordCluster}(object, max.level = 2L, ...)

\S4method{summary}{RecordCluster}(object)
}
\arguments{
\item{x}{\code{RecordCluster} object}

\item{object}{\code{RecordCluster} object}

\item{max.level}{Maximum level of nesting for str()}

\item{...}{Further arguments for str()}
}
\description{
Cluster record contains all information on a cluster.
}
\section{Slots}{

\describe{
\item{\code{id}}{Cluster ID, integer}

\item{\code{sids}}{Sequence IDs}

\item{\code{nsqs}}{Number of sequences}

\item{\code{txids}}{Source txids for sequences}

\item{\code{ntx}}{Number of taxa}

\item{\code{typ}}{Cluster type: direct, subtree or merged}

\item{\code{seed}}{Seed sequence ID}

\item{\code{prnt}}{Parent taxonomic ID}
}}

