% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes.R
\docType{class}
\name{RecordSequence-class}
\alias{RecordSequence-class}
\alias{RecordSequence-method}
\alias{as.character,RecordSequence-method}
\alias{show,RecordSequence-method}
\alias{print,RecordSequence-method}
\alias{str,RecordSequence-method}
\alias{summary,RecordSequence-method}
\title{RecordSequence-class}
\usage{
\S4method{as.character}{RecordSequence}(x)

\S4method{show}{RecordSequence}(object)

\S4method{print}{RecordSequence}(x)

\S4method{str}{RecordSequence}(object, max.level = 2L, ...)

\S4method{summary}{RecordSequence}(object)
}
\arguments{
\item{x}{\code{RecordSequence} object}

\item{object}{\code{RecordSequence} object}

\item{max.level}{Maximum level of nesting for str()}

\item{...}{Further arguments for str()}
}
\description{
Sequence record contains sequence data.
}
\details{
Sequence is stored as raw. Use rawToChar().
}
\section{Slots}{

\describe{
\item{\code{id}}{Unique ID}

\item{\code{nm}}{Best-guess sequence name}

\item{\code{accssn}}{Accession}

\item{\code{vrsn}}{Accession version}

\item{\code{url}}{URL}

\item{\code{gi}}{GI}

\item{\code{txid}}{Taxonomic ID of source taxon}

\item{\code{orgnsm}}{Scientific name of source taxon}

\item{\code{sq}}{Sequence}

\item{\code{dfln}}{Definition line}

\item{\code{ml_typ}}{Molecule type, e.g. DNA}

\item{\code{rcrd_typ}}{Record type: Whole or feature}

\item{\code{nncltds}}{Number of nucleotides}

\item{\code{nambgs}}{Number of ambiguous nucleotides}

\item{\code{pambgs}}{Proportion of ambiguous nucleotides}

\item{\code{gcr}}{GC ratio}

\item{\code{age}}{Number of days between sequence upload and running pipeline}
}}

