% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes.R
\docType{class}
\name{RecordTaxon-class}
\alias{RecordTaxon-class}
\alias{RecordTaxon-method}
\alias{as.character,RecordTaxon-method}
\alias{show,RecordTaxon-method}
\alias{print,RecordTaxon-method}
\alias{str,RecordTaxon-method}
\alias{summary,RecordTaxon-method}
\title{RecordTaxon-class}
\usage{
\S4method{as.character}{RecordTaxon}(x)

\S4method{show}{RecordTaxon}(object)

\S4method{print}{RecordTaxon}(x)

\S4method{str}{RecordTaxon}(object, max.level = 2L, ...)

\S4method{summary}{RecordTaxon}(object)
}
\arguments{
\item{x}{\code{RecordTaxon} object}

\item{object}{\code{RecordTaxon} object}

\item{max.level}{Maximum level of nesting for str()}

\item{...}{Further arguments for str()}
}
\description{
Taxonomic dictionary contains a taxonomic
tree and NCBI taxonomy data for all taxonomic IDs.
}
\section{Slots}{

\describe{
\item{\code{id}}{Taxonomic ID}

\item{\code{scnm}}{Scientific name}

\item{\code{cmnm}}{Common name}

\item{\code{rnk}}{Rank}

\item{\code{lng}}{Lineage}

\item{\code{prnt}}{Parent}
}}

