% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{remote.nodes.create}
\alias{remote.nodes.create}
\title{Create remote nodes}
\usage{
remote.nodes.create(root.taxa = c(33090, 4751, 33208),
  file.name = "nodes.tsv")
}
\description{
TODO
}
\examples{
# TODO
}
