% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{reset}
\alias{reset}
\title{Reset a PhyLoTa pipeline run}
\usage{
reset(wd, stage, hard = FALSE)
}
\arguments{
\item{wd}{Working directory}

\item{stage}{Name of stage to which the pipeline will be reset}

\item{hard}{T/F, delete all cached data?}
}
\description{
Resets the pipeline to a specified stage.
}
\seealso{
\code{\link{restart}}, \code{\link{setParameters}}
}
