% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{restart}
\alias{restart}
\title{Restart a PhyLoTa pipeline run}
\usage{
restart(wd)
}
\arguments{
\item{wd}{Working directory}
}
\description{
Restarts the running of a pipeline
as started with \code{run}.
}
\seealso{
\code{\link{run}}
}
