% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.R
\name{rmNdmtrx}
\alias{rmNdmtrx}
\title{Remove node matrix}
\usage{
rmNdmtrx(tree)
}
\arguments{
\item{tree}{\code{TreeMan} object}
}
\description{
Return tree with memory heavy node matrix removed.
}
\details{
Potential uses: reduce memory load of a tree,
save tree using serialization methods.
}
\examples{
#
tree <- randTree(10)
summary(tree)
tree <- rmNdmtrx(tree)
summary(tree)
}
\seealso{
\code{\link{addNdmtrx}}
}
