% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage4-tools.R
\name{rnmbrClstrs}
\alias{rnmbrClstrs}
\title{Renumbers cluster IDs}
\usage{
rnmbrClstrs(clstr_rcrds)
}
\arguments{
\item{clstr_rcrds}{List of clusters}
}
\description{
Returns a ClRcrdBx with
ID determined by the number of sequences
in each cluster.
}
