% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage4-tools.R
\name{rnmbrclusters}
\alias{rnmbrclusters}
\alias{clusters_renumber}
\title{Renumbers cluster IDs}
\usage{
clusters_renumber(cluster_rcrds)
}
\arguments{
\item{cluster_rcrds}{List of clusters}
}
\description{
Returns a ClRcrdBx with
ID determined by the number of sequences
in each cluster.
}
