% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-cache.R
\name{rstPrgrss}
\alias{rstPrgrss}
\title{Reset progress}
\usage{
rstPrgrss(wd, stg)
}
\arguments{
\item{wd}{Working directory}

\item{stg}{Stage to which the pipeline will be reset}
}
\description{
Reset progress to an earlier completed stage.
}
\details{
For example, resettting the progress to 'download'
mark stages 'download', 'cluster' and 'align' as un-run.
}
