% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{rstrtPhylota}
\alias{rstrtPhylota}
\title{Restart a PhyLoTa pipeline run}
\usage{
rstrtPhylota(wd)
}
\arguments{
\item{wd}{Working directory}
}
\description{
Restarts the running of a pipeline
as started with \code{runPhylota}.
}
\seealso{
\code{\link{runPhylota}}
}
