% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stages.R
\name{runAlign}
\alias{runAlign}
\title{Align clusters}
\usage{
runAlign(wd)
}
\arguments{
\item{wd}{Working directory}
}
\description{
Run external alignment software
on identified clusters.
}
\details{
Object will be cached.
}
