% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stages.R
\name{runClusters}
\alias{runClusters}
\title{Cluster}
\usage{
runClusters(wd)
}
\arguments{
\item{wd}{Working directory}
}
\description{
Identify sequence clusters
for an initial PhyLoTa cluster table
generated
}
