% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{runPhylota}
\alias{runPhylota}
\title{Run the entire PhyLoTa pipeline}
\usage{
runPhylota(wd, txid)
}
\arguments{
\item{wd}{Working directory}

\item{txid}{Root taxonomic ID}
}
\description{
Run the entire PhyLoTa pipeline.
All generated files will be stored in the wd.
The process can be stopped at anytime and 
restarted with \code{rstrtPhylota}.
}
\details{
All objects and data will be cached.
}
\seealso{
\code{\link{rstrtPhylota}}
}
