% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqs-tools.R
\name{rwRcrd2SqRcrd}
\alias{rwRcrd2SqRcrd}
\title{Convert raw Entrez gbwithparts record to SqRcrds}
\usage{
rwRcrd2SqRcrd(rw_rcrds, gis, ps)
}
\arguments{
\item{rw_rcrds}{Raw records returned from Entrez fetch}

\item{gis}{GIs of each fectched record}

\item{ps}{Parameters list}
}
\description{
Parses returned sequences features with Entrez,
returns a SqRcrd for each raw record.
}
