% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-api.R
\name{safely_connect}
\alias{safely_connect}
\title{Safely run rentrez function}
\usage{
safely_connect(func, args, fnm, ps)
}
\arguments{
\item{func}{rentrez function}

\item{args}{rentrez function arguments, list}

\item{fnm}{rentrez function name}

\item{ps}{Parameters}
}
\description{
Safely run a rentrez function.
If the query fails, the function will retry.
}
