% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-sqs.R
\name{seqrec_convert}
\alias{seqrec_convert}
\title{Convert raw Entrez gbwithparts record to SqRcrds}
\usage{
seqrec_convert(raw_recs, ps)
}
\arguments{
\item{raw_recs}{Raw records returned from Entrez fetch}

\item{ps}{Parameters list}

\item{gis}{GIs of each fectched record}
}
\description{
Parses returned sequences features with Entrez,
returns a SqRcrd for each raw record.
}
