% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage2-tools.R
\name{seqrec_get}
\alias{seqrec_get}
\title{seqrec_get}
\usage{
seqrec_get(txid, ps, drct = FALSE, lvl = 0)
}
\arguments{
\item{txid}{NCBI taxonomic ID}

\item{ps}{parameters}

\item{drct}{Node-level only or subtree as well? Default FALSE.}

\item{lvl}{Log level}
}
\value{
Vector of sequence records
}
\description{
Downloads sequences from GenBank in 500 ID batches.
}
