% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.R
\name{setAge}
\alias{setAge}
\title{Set the age of a tree}
\usage{
setAge(tree, val)
}
\arguments{
\item{tree}{\code{TreeMan} object}

\item{val}{new age}
}
\description{
Return a tree with the age altered.
}
\details{
Use this function to change the age of a tree. For example,
you might want to convert the tree so that its age equals 1. This function will achieve
that by modiyfing every branch, while maintaining their relative lengths.
}
\examples{

tree <- randTree(10)
tree <- setAge(tree, val = 1)
summary(tree)
}
\seealso{
\code{\link{setPD}}
\url{https://github.com/DomBennett/treeman/wiki/set-methods}
}
