% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{setup}
\alias{setup}
\title{Set-up parameters}
\usage{
setup(wd, txid, ncbi_dr = ".", v = FALSE, ...)
}
\arguments{
\item{wd}{Working directory}

\item{txid}{Root taxonomic ID(s), vector or numeric}

\item{ncbi_dr}{Directory to NCBI BLAST tools, default '.'}

\item{v}{Verbose, T/F}

\item{...}{Additional parameters}
}
\description{
Set up working directory with parameters.
}
\seealso{
\code{\link{parameters}}
}
