% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-cache.R
\name{setUpCch}
\alias{setUpCch}
\alias{cache_setup}
\title{Set-up a cache}
\usage{
cache_setup(ps, ovrwrt = FALSE)
}
\arguments{
\item{ps}{Paramters list generated by \code{setUpPrmtrs}.}

\item{ovrwrt}{Overwrite existing cache? Default FALSE.}
}
\description{
Creates a cache of parameters in the wd.
}
\details{
Warning: overwriting with this function will
delete the existing cache.
}
