% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{setUpNcbiTools}
\alias{setUpNcbiTools}
\title{Ensures NCBI BLAST tools are installed}
\usage{
setUpNcbiTools(dr)
}
\arguments{
\item{dr}{Directory to NCBI BLAST tools}
}
\description{
Ensures NCBI BLAST executables are installed on the 
system. Tests version number of BLAST tools.
}
\details{
BLAST tools must be version >2.7.1.
}
\seealso{
\link{mkPrmtrs}
}
