% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-cache.R
\name{sids_check}
\alias{sids_check}
\title{Check if sids exist}
\usage{
sids_check(wd, txid)
}
\arguments{
\item{wd}{Working directory}

\item{txid}{Taxonomic ID, numeric}
}
\value{
T/F
}
\description{
Check if sids are already downloaded for a txid.
}
\seealso{
Other run-private: 
\code{\link{batcher}()},
\code{\link{blast_clstr}()},
\code{\link{blast_filter}()},
\code{\link{blast_setup}()},
\code{\link{blast_sqs}()},
\code{\link{blastcache_load}()},
\code{\link{blastcache_save}()},
\code{\link{blastdb_gen}()},
\code{\link{blastn_run}()},
\code{\link{cache_rm}()},
\code{\link{cache_setup}()},
\code{\link{clade_select}()},
\code{\link{clstr2_calc}()},
\code{\link{clstr_all}()},
\code{\link{clstr_direct}()},
\code{\link{clstr_sqs}()},
\code{\link{clstr_subtree}()},
\code{\link{clstrarc_gen}()},
\code{\link{clstrarc_join}()},
\code{\link{clstrrec_gen}()},
\code{\link{clstrs_calc}()},
\code{\link{clstrs_join}()},
\code{\link{clstrs_merge}()},
\code{\link{clstrs_renumber}()},
\code{\link{clstrs_save}()},
\code{\link{cmdln}()},
\code{\link{descendants_get}()},
\code{\link{download_obj_check}()},
\code{\link{error}()},
\code{\link{gb_extract}()},
\code{\link{hierarchic_download}()},
\code{\link{info}()},
\code{\link{ncbicache_load}()},
\code{\link{ncbicache_save}()},
\code{\link{obj_check}()},
\code{\link{obj_load}()},
\code{\link{obj_save}()},
\code{\link{outfmt_get}()},
\code{\link{parameters_load}()},
\code{\link{parameters_setup}()},
\code{\link{parent_get}()},
\code{\link{progress_init}()},
\code{\link{progress_read}()},
\code{\link{progress_reset}()},
\code{\link{progress_save}()},
\code{\link{rank_get}()},
\code{\link{rawseqrec_breakdown}()},
\code{\link{safely_connect}()},
\code{\link{search_and_cache}()},
\code{\link{searchterm_gen}()},
\code{\link{seeds_blast}()},
\code{\link{seq_download}()},
\code{\link{seqarc_gen}()},
\code{\link{seqrec_augment}()},
\code{\link{seqrec_convert}()},
\code{\link{seqrec_gen}()},
\code{\link{seqrec_get}()},
\code{\link{sids_get}()},
\code{\link{sids_load}()},
\code{\link{sids_save}()},
\code{\link{sqs_count}()},
\code{\link{sqs_save}()},
\code{\link{stage_args_check}()},
\code{\link{stages_run}()},
\code{\link{tax_download}()},
\code{\link{taxdict_gen}()},
\code{\link{taxtree_gen}()},
\code{\link{txids_get}()},
\code{\link{txnds_count}()},
\code{\link{warn}()}
}
\concept{run-private}
