% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-sqrcrd.R
\docType{class}
\name{SqRcrd-class}
\alias{SqRcrd-class}
\alias{SqRcrd-method}
\alias{as.character,SqRcrd-method}
\alias{show,SqRcrd-method}
\alias{print,SqRcrd-method}
\alias{str,SqRcrd-method}
\alias{summary,SqRcrd-method}
\title{SqRcrd-class}
\usage{
\S4method{as.character}{SqRcrd}(x)

\S4method{show}{SqRcrd}(object)

\S4method{print}{SqRcrd}(x)

\S4method{str}{SqRcrd}(object, max.level = 2L, ...)

\S4method{summary}{SqRcrd}(object)
}
\arguments{
\item{x}{\code{SqRcrd} object}

\item{object}{\code{SqRcrd} object}
}
\description{
Sequence record contains sequence data.
}
\details{
Sequence is stored as raw. Use rawToChar().
}
\section{Slots}{

\describe{
\item{\code{id}}{Unique ID}

\item{\code{nm}}{Best-guess sequence name}

\item{\code{accssn}}{Accession}

\item{\code{vrsn}}{Accession version}

\item{\code{url}}{URL}

\item{\code{gi}}{GI}

\item{\code{txid}}{Taxonomic ID}

\item{\code{sq}}{Sequence}

\item{\code{dfln}}{Definition line}

\item{\code{ml_typ}}{Molecule type, e.g. DNA}

\item{\code{rcrd_typ}}{Record type: Whole or feature}

\item{\code{nncltds}}{Number of nucleotides}

\item{\code{nambgs}}{Number of ambiguous nucleotides}

\item{\code{pambgs}}{Proportion of ambiguous nucleotides}

\item{\code{gcr}}{GC ratio}

\item{\code{age}}{Number of days between sequence upload and running pipeline}
}}

\seealso{
\code{\link{genSqRcrd}}
}
