% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes.R
\docType{class}
\name{SqRcrdBx-class}
\alias{SqRcrdBx-class}
\alias{SqRcrdBx-method}
\alias{as.character,SqRcrdBx-method}
\alias{show,SqRcrdBx-method}
\alias{print,SqRcrdBx-method}
\alias{str,SqRcrdBx-method}
\alias{summary,SqRcrdBx-method}
\alias{[[,SqRcrdBx,character-method}
\alias{[,SqRcrdBx,character,missing,missing-method}
\title{SqRcrdBx-class}
\usage{
\S4method{as.character}{SqRcrdBx}(x)

\S4method{show}{SqRcrdBx}(object)

\S4method{print}{SqRcrdBx}(x)

\S4method{str}{SqRcrdBx}(object, max.level = 2L, ...)

\S4method{summary}{SqRcrdBx}(object)

\S4method{[[}{SqRcrdBx,character}(x, i)

\S4method{[}{SqRcrdBx,character,missing,missing}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{\code{SqRcrdBx} object}

\item{object}{\code{SqRcrdBx} object}

\item{max.level}{Maximum level of nesting for str()}

\item{...}{Further arguments for str()}

\item{i}{sid(s)}

\item{j}{Unused}

\item{drop}{Unused}
}
\description{
Multiple sequence records containing sequence data.
}
\details{
Sequences are stored as raw. Use rawToChar().
}
\section{Slots}{

\describe{
\item{\code{ids}}{Vector of Sequence Record IDs}

\item{\code{nncltds}}{Vector of sequence lengths}

\item{\code{nambgs}}{Vector of number of ambiguous nucleotides}

\item{\code{txids}}{Vector source txid associated with each sequence}

\item{\code{sqs}}{List of SqRcrds named by ID}
}}

