% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqsrcrd-dclrtn.R
\docType{class}
\name{SqsRcrd-class}
\alias{SqsRcrd-class}
\alias{SqsRcrd-method}
\alias{as.character,SqsRcrd-method}
\alias{show,SqsRcrd-method}
\alias{print,SqsRcrd-method}
\alias{str,SqsRcrd-method}
\alias{summary,SqsRcrd-method}
\title{SqsRcrd-class}
\usage{
\S4method{as.character}{SqsRcrd}(x)

\S4method{show}{SqsRcrd}(object)

\S4method{print}{SqsRcrd}(x)

\S4method{str}{SqsRcrd}(object, max.level = 2L, ...)

\S4method{summary}{SqsRcrd}(object)
}
\arguments{
\item{x}{\code{SqsRcrd} object}

\item{object}{\code{SqsRcrd} object}
}
\description{
Multiple sequence records containing sequence data.
}
\details{
Sequences are stored as raw. Use rawToChar().
}
\section{Slots}{

\describe{
\item{\code{ids}}{Vector of SqsRcrd IDs}

\item{\code{nncltds}}{Vector of sequence lengths}

\item{\code{nambgs}}{Vector of number of ambiguous nucleotides}

\item{\code{txids}}{Vector source txid associated with each sequence}

\item{\code{sqs}}{List of SqsRcrds named by ID}
}}

\seealso{
\code{\link{genSqsRcrd}}
}
