% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline-tools.R
\name{stages_run}
\alias{stages_run}
\title{Sequentially run each stage}
\usage{
stages_run(wd, to, frm, stgs_msg, rstrt = FALSE)
}
\arguments{
\item{wd}{Working directory}

\item{to}{Total number of stages to run}

\item{frm}{Starting stage to run from}

\item{stgs_msg}{Printout stage message for log}

\item{rstrt}{Restarting, T/F}
}
\description{
Runs stages from \code{frm} to \code{to}.
Records stage progress in cache.
}
