\name{sturgeons}
\alias{sturgeons}
\docType{data}
\title{Example phylota object (sturgeons)}
\description{
  Final phylota object produced by running pipeline for the sturgeons, Acipenseridae (7900).
}
\usage{
   data(sturgeons)
}
\format{
  \code{sturgeons} is generated from read_phylota().
}
\examples{
data(sturgeons) # load object
str(sturgeons)
}
\keyword{datasets}