% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache-tools.R
\name{svBlstCch}
\alias{svBlstCch}
\title{Save BLAST results to cache}
\usage{
svBlstCch(sqs, wd, obj)
}
\arguments{
\item{sqs}{Sequences}

\item{wd}{Working dir}

\item{obj}{BLAST result}
}
\description{
Run whenever local BLAST runs are made to save
results in cache in case the pipeline is run again.
}
