% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-cache.R
\name{svClstrs}
\alias{svClstrs}
\title{Save clusters to cache}
\usage{
svClstrs(wd, txid, clstrs)
}
\arguments{
\item{wd}{Working directory}

\item{txid}{Taxonomic ID, numeric}

\item{clstr}{cluster list}
}
\description{
Saves clusters generated by clstrAll to cache.
}
