% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-cache.R
\name{svclusters}
\alias{svclusters}
\alias{clusters_save}
\title{Save clusters to cache}
\usage{
clusters_save(wd, txid, clusters)
}
\arguments{
\item{wd}{Working directory}

\item{txid}{Taxonomic ID, numeric}

\item{clusters}{cluster list}
}
\description{
Saves clusters generated by clusterAll to cache.
}
