% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-cache.R
\name{svNcbiCch}
\alias{svNcbiCch}
\alias{ncbicache_save}
\title{Save NCBI query result to cache}
\usage{
ncbicache_save(fnm, args, wd, obj)
}
\arguments{
\item{fnm}{NCBI Entrez function name}

\item{args}{Args used for function}

\item{wd}{Working directory}

\item{obj}{NCBI query result}
}
\description{
Run whenever NCBI queries are made to save
results in cache in case the pipeline is run again.
}
