% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-cache.R
\name{svObj}
\alias{svObj}
\title{Aave a named object in the cache}
\usage{
svObj(wd, obj, nm)
}
\arguments{
\item{wd}{Working directory}

\item{obj}{Object}

\item{nm}{Object name}
}
\description{
Save an object in the cache
that can be loaded by \code{ldObj}
}
