% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-cache.R
\name{svPrgrss}
\alias{svPrgrss}
\alias{progress_save}
\title{Save current progress}
\usage{
progress_save(wd, stg)
}
\arguments{
\item{wd}{Working directory}

\item{stg}{Stage}
}
\description{
Stores the pipeline progress in the cache.
}
