% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-cache.R
\name{svSqs}
\alias{svSqs}
\alias{sqs_save}
\title{Save sequences to cache}
\usage{
sqs_save(wd, txid, sqs)
}
\arguments{
\item{wd}{Working directory}

\item{txid}{Taxonomic ID, numeric}

\item{sqs}{Sequences}
}
\description{
Saves sequences downloaded
}
\details{
Used within the \code{dwnld} function. Saves
sequence data by txid in cache.
}
