\name{tardigrades}
\alias{tardigrades}
\docType{data}
\title{Example phylota object (tardigrades)}
\description{
  Final phylota object produced by running pipeline for the tardigrades, Eutardigrada (42242).
}
\usage{
   data(tardigrades)
}
\format{
  \code{tardigrades} is generated from read_phylota().
}
\examples{
data(tardigrades) # load object
str(tardigrades)
}
\keyword{datasets}