% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage1-tools.R
\name{tax_download}
\alias{tax_download}
\title{Download Taxonomic Records}
\usage{
tax_download(ids, ps)
}
\arguments{
\item{ids}{Vector of taxonomic IDs}

\item{ps}{Parameter list}
}
\value{
list of list
}
\description{
Downloads one batch of taxonomic
records.
}
