% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes.R
\docType{class}
\name{TaxDict-class}
\alias{TaxDict-class}
\alias{TaxDict-method}
\alias{as.character,TaxDict-method}
\alias{show,TaxDict-method}
\alias{print,TaxDict-method}
\alias{str,TaxDict-method}
\alias{summary,TaxDict-method}
\title{TaxDict-class}
\usage{
\S4method{as.character}{TaxDict}(x)

\S4method{show}{TaxDict}(object)

\S4method{print}{TaxDict}(x)

\S4method{str}{TaxDict}(object, max.level = 2L, ...)

\S4method{summary}{TaxDict}(object)
}
\arguments{
\item{x}{\code{TaxDict} object}

\item{object}{\code{TaxDict} object}

\item{max.level}{Maximum level of nesting for str()}

\item{...}{Further arguments for str()}
}
\description{
Taxonomic dictionary contains a taxonomic
tree and NCBI taxonomy data for all taxonomic IDs.
}
\section{Slots}{

\describe{
\item{\code{txids}}{Taxonomic IDs of taxon records}

\item{\code{rcrds}}{Environment of records}

\item{\code{prnt}}{Parent taxonomic ID}

\item{\code{txtr}}{Taxonomic tree}
}}

