% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage1.R
\name{taxdict_gen}
\alias{taxdict_gen}
\title{Generate taxonomic dictionary}
\usage{
taxdict_gen(txids, rcrds, ps)
}
\arguments{
\item{txids}{Vector of taxonomic IDs}

\item{rcrds}{List of taxonomic records}
}
\value{
TaxDct
}
\description{
Takes a vector of txids and a list
of taxonomic records and returns a taxonomic dictionary.
}
