\name{tinamous}
\alias{tinamous}
\docType{data}
\title{Example phylota object (tinamous)}
\description{
  Final phylota object produced by running pipeline for the tinamous, Tinamiformes (8802).
}
\usage{
   data(tinamous)
}
\format{
  \code{tinamous} is generated from read_phylota().
}
\examples{
data(tinamous) # load object
str(tinamous)
}
\keyword{datasets}