% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.R
\name{TreeMan-to-phylo}
\alias{TreeMan-to-phylo}
\title{Convert TreeMan to phylo}
\description{
Return ape's \code{phylo} from a \code{TreeMan}. This will
preserve node labels if they are different from the default labels (n#).
}
\examples{

library(ape)
tree <- randTree(10)
tree <- as(tree, "phylo")
}
\seealso{
\code{\link{phylo-to-TreeMan}},
\code{\link{TreeMen-to-multiPhylo}}
\code{\link{multiPhylo-to-TreeMen}}
\code{\link{TreeMan-class}}
}
