% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.R
\docType{class}
\name{TreeMen-class}
\alias{TreeMen-class}
\alias{TreeMen-method}
\alias{cTrees,TreeMen-method}
\alias{[[,TreeMen,ANY-method}
\alias{[,TreeMen,character,missing,missing-method}
\alias{Extract}
\alias{slots}
\alias{from}
\alias{a}
\alias{list}
\alias{of}
\alias{trees}
\alias{as.character,TreeMen-method}
\alias{show,TreeMen-method}
\alias{str,TreeMen-method}
\alias{print,TreeMen-method}
\alias{summary,TreeMen-method}
\title{TreeMen-class}
\usage{
\S4method{cTrees}{TreeMen}(x, ...)

\S4method{[[}{TreeMen,ANY}(x, i)

\S4method{[}{TreeMen,character,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{as.character}{TreeMen}(x)

\S4method{show}{TreeMen}(object)

\S4method{str}{TreeMen}(object, max.level = 2L, ...)

\S4method{print}{TreeMen}(x)

\S4method{summary}{TreeMen}(object)
}
\arguments{
\item{x}{\code{TreeMen} object}

\item{...}{additional tree objects}

\item{i}{tree index (integer or character)}

\item{j}{missing}

\item{drop}{missing}

\item{object}{\code{TreeMen} object}

\item{max.level}{\code{str()} maximum level}
}
\description{
S4 class for multiple phylogenetic trees
}
\section{Slots}{

\describe{
\item{\code{treelst}}{list of \code{TreeMan} objects}

\item{\code{ntips}}{sum of tips per tree}

\item{\code{ntrees}}{total number of trees}
}}

\seealso{
\code{\link{cTrees}}
}
