% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.R
\name{TreeMen-to-multiPhylo}
\alias{TreeMen-to-multiPhylo}
\title{Convert TreeMen to multiPhylo}
\description{
Return ape's \code{multiPhylo} from a \code{TreeMen}
}
\examples{

library(ape)
trees <- cTrees(randTree(10), randTree(10), randTree(10))
trees <- as(trees, "multiPhylo")
}
\seealso{
\code{\link{TreeMan-to-phylo}},
\code{\link{phylo-to-TreeMan}},
\code{\link{multiPhylo-to-TreeMen}}
\code{\link{TreeMan-class}}
}
