% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-tools-txdct.R
\name{trid2Txid}
\alias{trid2Txid}
\title{Get taxonomic ID from tree ID}
\usage{
trid2Txid(id, txdct)
}
\arguments{
\item{id}{txid}

\item{txdct}{TxDct}
}
\value{
Character or vector
}
\description{
Return taxonomic ID for taxonomic
tree ID in TxDct.
}
