% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.R
\name{twoer}
\alias{twoer}
\title{Generate a tree of two tips}
\usage{
twoer(tids = c("t1", "t2"), spns = c(1, 1), rid = "root", root_spn = 0)
}
\arguments{
\item{tids}{tip IDs}

\item{spns}{tip spans}

\item{rid}{root ID}

\item{root_spn}{root span}
}
\description{
Returns a \code{TreeMan} tree with two tips and a root.
}
\details{
Useful for building larger trees with \code{addClade()}.
Note, a node matrix cannot be added to a tree of two tips.
}
\examples{

tree <- twoer()
}
\seealso{
\code{\link{TreeMan-class}}, \code{\link{randTree}}
}
