% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-txdct.R
\docType{class}
\name{TxDct-class}
\alias{TxDct-class}
\alias{TxDct-method}
\alias{as.character,TxDct-method}
\alias{show,TxDct-method}
\alias{print,TxDct-method}
\alias{str,TxDct-method}
\alias{summary,TxDct-method}
\title{TxDct-class}
\usage{
\S4method{as.character}{TxDct}(x)

\S4method{show}{TxDct}(object)

\S4method{print}{TxDct}(x)

\S4method{str}{TxDct}(object, max.level = 2L, ...)

\S4method{summary}{TxDct}(object)
}
\arguments{
\item{x}{\code{TxDct} object}

\item{object}{\code{TxDct} object}
}
\description{
Cluster record contains all information on a cluster.
}
\section{Slots}{

\describe{
\item{\code{txids}}{Taxonomic IDs of taxon records}

\item{\code{rcrds}}{List of records}

\item{\code{prnt}}{Parent taxonomic ID}

\item{\code{txtr}}{Taxonomic tree}
}}

\seealso{
\code{\link{genTxDct}}
}
