% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-tools-txdct.R
\name{txid2Trid}
\alias{txid2Trid}
\title{Get tree ID from taxonomic ID}
\usage{
txid2Trid(id, txdct)
}
\arguments{
\item{id}{trid}

\item{txdct}{TxDct}
}
\value{
Character or vector
}
\description{
Return taxonomic tree ID from taxonomic
ID in TxDct.
}
