% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes.R
\docType{class}
\name{TxRcrd-class}
\alias{TxRcrd-class}
\alias{TxRcrd-method}
\alias{as.character,TxRcrd-method}
\alias{show,TxRcrd-method}
\alias{print,TxRcrd-method}
\alias{str,TxRcrd-method}
\alias{summary,TxRcrd-method}
\title{TxRcrd-class}
\usage{
\S4method{as.character}{TxRcrd}(x)

\S4method{show}{TxRcrd}(object)

\S4method{print}{TxRcrd}(x)

\S4method{str}{TxRcrd}(object, max.level = 2L, ...)

\S4method{summary}{TxRcrd}(object)
}
\arguments{
\item{x}{\code{TxRcrd} object}

\item{object}{\code{TxRcrd} object}

\item{max.level}{Maximum level of nesting for str()}

\item{...}{Further arguments for str()}
}
\description{
Taxonomic dictionary contains a taxonomic
tree and NCBI taxonomy data for all taxonomic IDs.
}
\section{Slots}{

\describe{
\item{\code{id}}{Taxonomic ID}

\item{\code{scnm}}{Scientific name}

\item{\code{cmnm}}{Common name}

\item{\code{rnk}}{Rank}

\item{\code{lng}}{Lineage}

\item{\code{prnt}}{Parent}
}}

