% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-tools-phylota-special.R
\name{update_phylota}
\alias{update_phylota}
\title{Update slots}
\usage{
update_phylota(phylota)
}
\arguments{
\item{phylota}{Phylota object}
}
\description{
After change, run to update
slots.
}
