% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.R
\name{updateSlts}
\alias{updateSlts}
\title{Update tree slots after manipulation}
\usage{
updateSlts(tree)
}
\arguments{
\item{tree}{\code{TreeMan} object}
}
\description{
Return tree with updated slots.
}
\details{
Tree slots in the \code{TreeMan} object are usually automatically updated.
For certain single node manipulations they are not. Run this
function to update the slots.
}
\seealso{
\code{\link{addNdmtrx}}, \code{\link{getAge}}
}
