% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-tools.R
\name{warn}
\alias{warn}
\title{Write warning message to log}
\usage{
warn(wd, ...)
}
\arguments{
\item{wd}{Working directory}

\item{...}{Message elements for concatenating}
}
\description{
Inform a user if a potential error has occurred in log.txt.
}
