% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-special.R
\name{write_sqs}
\alias{write_sqs}
\title{Write out sequences}
\usage{
write_sqs(phylota, outfile, sid, sq_nm = sid)
}
\arguments{
\item{phylota}{Phylota object}

\item{outfile}{Output file}

\item{sid}{Sequence ID(s)}

\item{sq_nm}{Sequence name(s)}
}
\description{
Write out sequences, as .fasta,
for a given vector of IDs.
}
\details{
The user can control the output definition
lines of the sequences using the sq_nm.
By default sequences IDs are used.
Note, ensure the sq_nm are in the same
order as sid.
}
