% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-tools.R
\name{writeClstr}
\alias{writeClstr}
\title{Write out PhyLoTa cluster}
\usage{
writeClstr(txid, cldf, cigidf, sqs, ps)
}
\arguments{
\item{phylt_nds}{PhyLoTa nodes data.frame}
}
\description{
Takes PhyLoTa data.frame and writes
as .tsv
}
\details{
PhyLoTa data.frame must be informed by
clustering functions before writing out.
}
