% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clstrsobj-tools.R
\name{writeSeqs}
\alias{writeSeqs}
\title{Write sequences as fasta}
\usage{
writeSeqs(clstrs_obj, id, wd)
}
\arguments{
\item{clstrs_obj}{Clusters Object}

\item{id}{unique cluster ID}

\item{wd}{Working directory, where files will be saved}
}
\description{
TODO
}
