% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylota-tools.R
\name{writeSqs}
\alias{writeSqs}
\title{Write sequences as fasta}
\usage{
writeSqs(sids, dflns, flpth)
}
\arguments{
\item{sids}{Sequence IDs}

\item{dflns}{Definition lines}

\item{flpth}{File path and name}
}
\description{
TODO
}
