\name{yeasts}
\alias{yeasts}
\docType{data}
\title{Example phylota object (yeasts)}
\description{
  Final phylota object produced by running pipeline for the yeasts, Kazachstania (71245).
}
\usage{
   data(yeasts)
}
\format{
  \code{yeasts} is generated from read_phylota().
}
\examples{
data(yeasts) # load object
str(yeasts)
}
\keyword{datasets}