% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage-tools-download.R
\name{agmntSqRcrds}
\alias{agmntSqRcrds}
\title{Augment sequence records list}
\usage{
agmntSqRcrds(sqs, txdct)
}
\arguments{
\item{sqs}{List of SqRcrds}

\item{txdct}{Taxonomic Dictionary}
}
\description{
Convert to sqsrcrds and add taxids
}
