% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-blast.R
\name{blstN}
\alias{blstN}
\title{BLAST all vs all}
\usage{
blstN(dbfl, outfl, ps)
}
\arguments{
\item{dbfl}{Database file}

\item{outfl}{Output file}
}
\description{
Use \code{blastn} to BLAST all vs all using
a BLAST database
}
