% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-blast.R
\name{blstN}
\alias{blstN}
\alias{blastn_run}
\title{BLAST all vs all}
\usage{
blastn_run(dbfl, outfl, ps)
}
\arguments{
\item{dbfl}{Database file}

\item{outfl}{Output file}

\item{ps}{Parameters}
}
\description{
Use \code{blastn} to BLAST all vs all using
a BLAST database
}
