% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage-tools-cluster2.R
\name{blstSeeds}
\alias{blstSeeds}
\title{BLAST seed sequences}
\usage{
blstSeeds(sqs, ps)
}
\arguments{
\item{sqs}{All seed sequences to be BLASTed}

\item{ps}{Parameters}
}
\description{
Runs all-v-all blast for seed sequences.
}
