% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage-tools-cluster.R
\name{blstSqs}
\alias{blstSqs}
\title{BLAST All vs All}
\usage{
blstSqs(txid, typ, sqs, ps, lvl)
}
\arguments{
\item{txid}{Taxonomic node ID, numeric}

\item{typ}{Cluster type, 'direct' or 'subtree'}

\item{sqs}{Sequences}

\item{ps}{Parameters}

\item{lvl}{Log level}
}
\description{
Return blast results from
BLASTing all vs all for given sequences
}
