% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage-download.R
\name{cldIdntfy}
\alias{cldIdntfy}
\title{Get all node IDs that will be processed}
\usage{
cldIdntfy(txdct, ps)
}
\arguments{
\item{txdct}{TxDct}

\item{ps}{Parameters}
}
\description{
All nodes for which all children
contain < mxsqs sequences
}
