% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes.R
\docType{class}
\name{ClRcrdBx-class}
\alias{ClRcrdBx-class}
\alias{ClRcrdBx-method}
\alias{as.character,ClRcrdBx-method}
\alias{show,ClRcrdBx-method}
\alias{print,ClRcrdBx-method}
\alias{str,ClRcrdBx-method}
\alias{summary,ClRcrdBx-method}
\alias{[[,ClRcrdBx,character-method}
\alias{[,ClRcrdBx,character,missing,missing-method}
\title{ClRcrdBx-class}
\usage{
\S4method{as.character}{ClRcrdBx}(x)

\S4method{show}{ClRcrdBx}(object)

\S4method{print}{ClRcrdBx}(x)

\S4method{str}{ClRcrdBx}(object, max.level = 2L, ...)

\S4method{summary}{ClRcrdBx}(object)

\S4method{[[}{ClRcrdBx,character}(x, i)

\S4method{[}{ClRcrdBx,character,missing,missing}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{\code{ClRcrdBx} object}

\item{object}{\code{ClRcrdBx} object}

\item{max.level}{Maximum level of nesting for str()}

\item{...}{Further arguments for str()}

\item{i}{cid(s)}

\item{j}{Unused}

\item{drop}{Unused}
}
\description{
Multiple cluster records.
}
\section{Slots}{

\describe{
\item{\code{ids}}{Vector of Cluster Record IDs}

\item{\code{cls}}{List of ClRcrds named by ID}
}}

