% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage-tools-cluster.R
\name{clstrAll}
\alias{clstrAll}
\title{Hierarchically cluster all sequences of a txid}
\usage{
clstrAll(txid, sqs, phylt_nds, ps, lvl = 0)
}
\arguments{
\item{txid}{Taxonomic ID}

\item{sqs}{Sequence object of all downloaded sequences}

\item{phylt_nds}{PhyLoTa table}

\item{ps}{Parameters}

\item{lvl}{Log level}
}
\description{
Identifies all direct and subtree clusters
for a taxonomic ID.
}
