% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage3-tools.R
\name{clstrBlstRs}
\alias{clstrBlstRs}
\title{Cluster BLAST Results}
\usage{
clstrBlstRs(blst_rs)
}
\arguments{
\item{blst_rs}{BLAST results}
}
\value{
List of list
}
\description{
Find single-linkage clusters from
BLAST results. Identifies seed sequence.
}
