% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage3-tools.R
\name{clstrDrct}
\alias{clstrDrct}
\title{Cluster sequences directly associated with txid}
\usage{
clstrDrct(txid, sqs, txdct, ps, lvl)
}
\arguments{
\item{txid}{Taxonomic ID}

\item{sqs}{Sequence object of all downloaded sequences}

\item{txdct}{PhyLoTa table}

\item{ps}{Parameters}

\item{lvl}{Log level}
}
\description{
In GenBank certain sequences may only be associated
with a higher level taxon (e.g. genus, family ...). This function
generates clusters from these sequences, alone. This function
identifies such sequences in the sequence object and generates
a list of clusters of cl_type 'Node'.
}
