% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-tools.R
\name{clstrSbtr}
\alias{clstrSbtr}
\title{Cluster all sequences descending from a txid}
\usage{
clstrSbtr(txid, sqs, phylt_nds, dds, ps, lvl)
}
\arguments{
\item{txid}{Taxonomic ID}

\item{sqs}{Sequence object of all downloaded sequences}

\item{phylt_nds}{PhyLoTa table}

\item{dds}{Vector of direct descendants}

\item{ps}{Parameters}
}
\description{
Identifies clusters from sequences associated
with a txid and all its descendants. Clusters returned by
this function will thus be of cl_type 'subtree'.
}
