% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clstrsobj-declaration.R
\docType{class}
\name{ClstrsObj-class}
\alias{ClstrsObj-class}
\alias{ClstrsObj-method}
\alias{as.character,ClstrsObj-method}
\alias{show,ClstrsObj-method}
\alias{print,ClstrsObj-method}
\alias{str,ClstrsObj-method}
\alias{summary,ClstrsObj-method}
\title{ClstrsObj-class}
\usage{
\S4method{as.character}{ClstrsObj}(x)

\S4method{show}{ClstrsObj}(object)

\S4method{print}{ClstrsObj}(x)

\S4method{str}{ClstrsObj}(object, max.level = 2L, ...)

\S4method{summary}{ClstrsObj}(object)
}
\arguments{
\item{x}{\code{ClstrsObj} object}

\item{object}{\code{ClstrsObj} object}
}
\description{
Clusters Object contains sequences and clusters information.
}
\section{Slots}{

\describe{
\item{\code{sqs}}{All sequences}

\item{\code{clstrs}}{All clstrs}
}}

\seealso{
\code{\link{genClstrsObj}}
}
