% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-tools.R
\name{clstrSqs}
\alias{clstrSqs}
\title{Cluster sequences for sequences}
\usage{
clstrSqs(wd, txid, sqs, phylt_nds, drct = FALSE, infrmtv = FALSE,
  blst_rs = NULL, verbose = FALSE)
}
\arguments{
\item{phylt_nds}{PhyLoTa nodes data.frame}
}
\description{
Identify sequence clusters using BLAST
}
