% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-tools.R
\name{clstrSqs}
\alias{clstrSqs}
\title{Identify clusters from sequences}
\usage{
clstrSqs(txid, sqs, phylt_nds, ps, lvl, typ = c("direct", "subtree"))
}
\arguments{
\item{txid}{Taxonomic ID}

\item{sqs}{Sequence object of sequences to be BLASTed}

\item{phylt_nds}{PhyLoTa table}

\item{ps}{Parameters}

\item{typ}{Direct or Subtree?}
}
\description{
Given a sequence object, this function will generate
PhyLoTa-like tables using BLAST.
}
