% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage3.R
\name{clusters_run}
\alias{clusters_run}
\title{Cluster}
\usage{
clusters_run(wd)
}
\arguments{
\item{wd}{Working directory}
}
\description{
Identify sequence clusters
for an initial PhyLoTa cluster table
generated
}
