% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-system.R
\name{cmdln}
\alias{cmdln}
\title{Run a command via terminal/command prompt}
\usage{
cmdln(cmd, args, lgfl = NULL)
}
\arguments{
\item{cmd}{Command to be run}

\item{args}{Vector of command arguments, each parameter and value
must be a separate element}

\item{lgfl}{File to which stdout/err will be written}
}
\value{
status, integer or character
}
\description{
Provide the command and arguments as a vector.
Also can take a lgfl to which all stdout and stderr is written.
If lgfl is not provided, a list is returned of 'status', 'stdout'
and 'stderr'. Else only the status is returned - 1 success, 0
failed.
}
\details{
Note, stdout/err are returned as 'raw'. Use rawToChar() to
convert to characters.
}
\seealso{
Other run-private: \code{\link{batcher}},
  \code{\link{blast_clstr}}, \code{\link{blast_filter}},
  \code{\link{blast_setup}}, \code{\link{blast_sqs}},
  \code{\link{blastcache_load}},
  \code{\link{blastcache_save}}, \code{\link{blastdb_gen}},
  \code{\link{blastn_run}}, \code{\link{cache_rm}},
  \code{\link{cache_setup}}, \code{\link{clade_select}},
  \code{\link{clstr2_calc}}, \code{\link{clstr_all}},
  \code{\link{clstr_direct}}, \code{\link{clstr_sqs}},
  \code{\link{clstr_subtree}}, \code{\link{clstrarc_gen}},
  \code{\link{clstrarc_join}}, \code{\link{clstrrec_gen}},
  \code{\link{clstrs_calc}}, \code{\link{clstrs_join}},
  \code{\link{clstrs_merge}},
  \code{\link{clstrs_renumber}}, \code{\link{clstrs_save}},
  \code{\link{descendants_get}},
  \code{\link{download_obj_check}}, \code{\link{error}},
  \code{\link{hierarchic_download}}, \code{\link{info}},
  \code{\link{ncbicache_load}},
  \code{\link{ncbicache_save}}, \code{\link{obj_check}},
  \code{\link{obj_load}}, \code{\link{obj_save}},
  \code{\link{parameters_load}},
  \code{\link{parameters_setup}}, \code{\link{parent_get}},
  \code{\link{progress_init}}, \code{\link{progress_read}},
  \code{\link{progress_reset}},
  \code{\link{progress_save}}, \code{\link{rank_get}},
  \code{\link{safely_connect}},
  \code{\link{search_and_cache}},
  \code{\link{searchterm_gen}}, \code{\link{seeds_blast}},
  \code{\link{seq_download}}, \code{\link{seqarc_gen}},
  \code{\link{seqrec_augment}},
  \code{\link{seqrec_convert}}, \code{\link{seqrec_gen}},
  \code{\link{seqrec_get}}, \code{\link{sids_check}},
  \code{\link{sids_get}}, \code{\link{sids_load}},
  \code{\link{sids_save}}, \code{\link{sqs_count}},
  \code{\link{sqs_load}}, \code{\link{sqs_save}},
  \code{\link{stage_args_check}}, \code{\link{stages_run}},
  \code{\link{tax_download}}, \code{\link{taxdict_gen}},
  \code{\link{taxtree_gen}}, \code{\link{txids_get}},
  \code{\link{txnds_count}}, \code{\link{warn}}
}
