% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-drop.R
\name{drop_by_rank}
\alias{drop_by_rank}
\title{Reduce clusters to specific rank}
\usage{
drop_by_rank(phylota, rnk = "species", keep_higher = FALSE, n = 10,
  choose_by = c("pambgs", "age", "nncltds"), greatest = c(FALSE, FALSE,
  TRUE))
}
\arguments{
\item{phylota}{Phylota object}

\item{rnk}{Taxonomic rank}

\item{keep_higher}{Keep higher taxonomic ranks?}

\item{n}{Number of sequences per taxon}

\item{choose_by}{Vector of selection functions}

\item{greatest}{Greatest of lowest for each choose_by function}
}
\value{
phylota
}
\description{
Identifies higher level taxa for each sequence in clusters for
given rank. Selects representative sequences for each unique taxon using the
choose_by functions. By default, the function will choose the top ten
sequences by first sorting by those with fewest number of ambiguous
sequences, then by youngest, then by sequence length.
}
\seealso{
Other tools-public: \code{\link{calc_mad}},
  \code{\link{calc_wrdfrq}}, \code{\link{drop_clstrs}},
  \code{\link{drop_sqs}}, \code{\link{get_clstr_slot}},
  \code{\link{get_nsqs}}, \code{\link{get_ntaxa}},
  \code{\link{get_sq_slot}}, \code{\link{get_stage_times}},
  \code{\link{get_tx_slot}}, \code{\link{get_txids}},
  \code{\link{is_txid_in_clstr}},
  \code{\link{is_txid_in_sq}},
  \code{\link{list_clstrrec_slots}},
  \code{\link{list_ncbi_ranks}},
  \code{\link{list_seqrec_slots}},
  \code{\link{list_taxrec_slots}},
  \code{\link{plot_phylota_pa}},
  \code{\link{plot_phylota_treemap}},
  \code{\link{read_phylota}}, \code{\link{write_sqs}}
}
