% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage-download.R
\name{dwnld}
\alias{dwnld}
\title{Download sequences for txids}
\usage{
dwnld(txids, phylt_nds, txdct, ps)
}
\arguments{
\item{txids}{Taxonomic node IDs, numeric vector}

\item{phylt_nds}{PhyLoTa data.frame}

\item{txdct}{Taxonomic dictionary}

\item{verbose}{Verbose? T/F}
}
\description{
Look up and download all sequences for
given taxonomic IDs.
}
\details{
Sequence downloads are cached.
}
