% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi-tools.R
\name{dwnldFrmNCBI}
\alias{dwnldFrmNCBI}
\title{dwnldFrmNCBI}
\usage{
dwnldFrmNCBI(txid, ps, direct = FALSE)
}
\arguments{
\item{txid}{NCBI taxon identifier}

\item{direct}{Whether to download all 'directly' linked
sequences ([Organism:noexp] in Genbank search) or all 'subtree'
sequences ([Organism:exp])
seqs exceeds it.
Only makes sense when direct=TRUE}
}
\value{
list of lists containing sequence objects
}
\description{
Given a taxon ID, queries NCBI
for sequences that match the taxid.
}
