% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage-tools-taxise.R
\name{dwnldTD}
\alias{dwnldTD}
\title{Download NCBI taxonomy dump}
\usage{
dwnldTD(ps)
}
\arguments{
\item{ps}{Parameters}
}
\description{
Download and unpack NCBI
taxonomy dump.
}
\details{
If \code{tdpth} is left NULL, downloads
the taxdump.tar.gz. Otherwise, manually download
the file and provide path.
\url{ftp://ftp.ncbi.nlm.nih.gov/pub/taxonomy/taxdump.tar.gz}
}
