% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage-taxise.R
\name{dwnldTxRcrds}
\alias{dwnldTxRcrds}
\title{Download taxonomic records}
\usage{
dwnldTxRcrds(txids, ps)
}
\arguments{
\item{txids}{Vector of taxonomic IDs}

\item{ps}{Parameter list}
}
\value{
List of taxonomic records
}
\description{
Takes a vector of txids and returns
a list of taxonomic records.
}
