% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-log.R
\name{error}
\alias{error}
\title{Write error message to log}
\usage{
error(ps, ...)
}
\arguments{
\item{ps}{parameters}

\item{...}{Message elements for concatenating}
}
\description{
Inform a user if an error has occurred in log.txt,
halt pipeline.
}
