% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-tools.R
\name{error}
\alias{error}
\title{Write error message to log}
\usage{
error(ps, ...)
}
\arguments{
\item{...}{Message elements for concatenating}

\item{wd}{Working directory}
}
\description{
Inform a user if an error has occurred in log.txt,
halt pipeline.
}
