% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-blast.R
\name{fltrBlstRs}
\alias{fltrBlstRs}
\alias{blast_filter}
\title{Filter BLAST results}
\usage{
blast_filter(blast_res, ps)
}
\arguments{
\item{blast_res}{BLAST results}

\item{ps}{Parameters}
}
\description{
Given a BLAST output, filters query-subject pairs such
that only HSPs with a coverage greater than \code{mncvrg} (specified
in the pipeline parameters) remain. Filters both: query-subject and
subject-query pairs, if one of the coverages is insufficient. HSP
coverage is obtained from the BLAST column \code{qcovs}.
}
