% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxise-tools.R
\name{genPhylotaNds}
\alias{genPhylotaNds}
\title{Generate taxonomic nodes in PhyLoTa format}
\usage{
genPhylotaNds(nids, verbose = FALSE)
}
\arguments{
\item{nids}{Taxonomic node IDs, character vector}

\item{verbose}{Verbose? T/F}
}
