% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxise-tools.R
\name{genPhylotaNds}
\alias{genPhylotaNds}
\title{Generate taxonomic nodes in PhyLoTa format}
\usage{
genPhylotaNds(nid_sets, mx_sq_lngth, mdl_thrshld, td_nds, td_nms, verbose)
}
\arguments{
\item{nid_sets}{Taxonomic processable and unprocessable
node IDs, list as generated by \code{getMngblIds}}

\item{verbose}{Verbose? T/F}
}
