% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage-taxise.R
\name{genTDObj}
\alias{genTDObj}
\title{Generate R object from NCBI taxonomy dump}
\usage{
genTDObj(ps)
}
\arguments{
\item{ps}{Parameters}
}
\description{
Generates an interrogable R object
from the NCBI taxonomy dump using the \code{CHNOSZ}
library. The returned object is a list of taxonomic nodes
and names.
}
\details{
Object will be cached.
}
