% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage-taxise.R
\name{genTxdct}
\alias{genTxdct}
\title{Generate taxonomic dictionary}
\usage{
genTxdct(phylt_nds, ps)
}
\arguments{
\item{phylt_nds}{PhyLoTa nodes as generated by \code{genPhylotaNds}}
}
\description{
Look-up all taxonomic data available in NCBI for
IDs in PhyLoTa table. Returns a list of esummary objects.
}
