% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage-tools-cluster.R
\name{getADs}
\alias{getADs}
\title{Get all descendants}
\usage{
getADs(txid, phylt_nds)
}
\arguments{
\item{txid}{Taxonomic node ID, numeric}

\item{phylt_nds}{PhyLoTa nodes data.frame}
}
\description{
Return all the taxonomic node IDs descending
from given taxonomic ID
}
