% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxise-tools.R
\name{getMngblIds}
\alias{getMngblIds}
\title{Identify manageable taxonomic node IDs}
\usage{
getMngblIds(txid, td_nds, mx_dscndnts = 10000, tmout = 10,
  verbose = FALSE)
}
\arguments{
\item{txid}{Root taxonomic ID(s), vector or character}

\item{td_nds}{'nds' element from the \code{tdobj}}

\item{mx_dscndnts}{Maximum number of descendants for
'manageable' taxonomic node, numeric.}

\item{tmout}{Number of seconds before discarding node,
numeric.}

\item{verbose}{print progress to screen? T/F}
}
\description{
Given a root \code{txid}, return a set of
taxonomic node IDs that only have up to a maximum number
of descendants.
}
\details{
Returns a \code{list}.
If there are more descendants than mx_dscndnts or
retrieveing the number takes longer than \code{tmout}
seconds, the taxonomic ID is discarded and its children
are added to the queue.
}
