% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-tools-txdct.R
\name{getPrnt}
\alias{getPrnt}
\title{Get taxonomic parent}
\usage{
getPrnt(id, txdct)
}
\arguments{
\item{id}{txid(s)}

\item{txdct}{TxDct}
}
\value{
Character
}
\description{
Look-up MRCA of id(s) from TxDct
}
