% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-co-sq.R
\name{getSqDfs}
\alias{getSqDfs}
\title{Return sequence deflines}
\usage{
getSqDfs(clstrs_obj, id, prse = 0.1)
}
\arguments{
\item{clstrs_obj}{clstrs_obj}

\item{id}{cluster ID(s)}
}
\description{
Return all deflines for a cluster's sequences
}
