% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co-sq-tools.R
\name{getSqLns}
\alias{getSqLns}
\title{Return sequence deflines}
\usage{
getSqLns(clstrs_obj, cid = NULL, sid = NULL)
}
\arguments{
\item{clstrs_obj}{clstrs_obj}

\item{id}{cluster ID(s)}
}
\description{
Return all deflines for a cluster's sequences
}
