% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage-tools-download.R
\name{getSqsByTxid}
\alias{getSqsByTxid}
\title{Hierarchically get sequences for a txid}
\usage{
getSqsByTxid(txid, txdct, ps, lvl = 0)
}
\arguments{
\item{txid}{Taxonomic node ID, numeric}
}
\description{
Looks up and downloads sequences for a
taxonomic ID.
}
