% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-tools.R
\name{getSqsByTxid}
\alias{getSqsByTxid}
\title{Hierarchically get sequences for a txid}
\usage{
getSqsByTxid(txid, phylt_nds, ps)
}
\arguments{
\item{txid}{Taxonomic node ID, numeric}

\item{phylt_nds}{PhyLoTa data.frame}

\item{verbose}{Verbose? T/F}
}
\description{
Looks up and downloads sequences for a
taxonomic ID.
}
