% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxise-tools.R
\name{getStats}
\alias{getStats}
\title{Generate PhyLoTa statistics for txid}
\usage{
getStats(wd, txid, phylt_nds, td_nds, td_nms, mx_sq_lngth, mdl_thrshld,
  verbose = FALSE, recursive = FALSE)
}
\arguments{
\item{txid}{Root taxonomic ID(s), vector or character}

\item{phylt_nds}{PhyLoTa data.frame}

\item{td_nds}{'nds' element from the \code{tdobj}}

\item{td_nms}{'nms' element from the \code{tdobj}}

\item{mx_sq_lngth}{Maximum number of associated sequences, numeric}

\item{verbose}{Print progress to screen? T/F}

\item{recursive}{Add stats for kid nodes recursively? T/F}

\item{mdl_threshold}{Maximum threshold, numeric.}
}
\details{
Returns an updated \code{phylt_nds}.
}
