% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage2-tools.R
\name{hrrchcDwnld}
\alias{hrrchcDwnld}
\title{Hierarchically get sequences for a txid}
\usage{
hrrchcDwnld(txid, txdct, ps, lvl = 0)
}
\arguments{
\item{txid}{Taxonomic node ID, numeric}

\item{txdct}{Taxonomic dictionary}

\item{ps}{parameters}

\item{lvl}{Log level}
}
\description{
Looks up and downloads sequences for a
taxonomic ID.
}
