% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-tools.R
\name{info}
\alias{info}
\title{Write info message to log}
\usage{
info(lvl, ps, ...)
}
\arguments{
\item{lvl}{Number of message indentations, indicating code depth.}

\item{ps}{Parameters}

\item{...}{Message elements for concatenating}
}
\description{
Inform a user via log.txt of pipeline
progress.
}
