% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-cache.R
\name{intPrgrss}
\alias{intPrgrss}
\alias{progress_init}
\title{Initialise progress list in cache}
\usage{
progress_init(wd)
}
\arguments{
\item{wd}{Working directory}
}
\description{
Creates a progress list recording each
stage run in cache
}
