% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-cache.R
\name{ldBlstCch}
\alias{ldBlstCch}
\title{Load BLAST results from cache}
\usage{
ldBlstCch(sids, wd)
}
\arguments{
\item{sids}{Sequence IDs}

\item{wd}{Working dir}

\item{obj}{BLAST result}
}
\description{
Run to load cached BLAST results.
}
