% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache-tools.R
\name{ldNcbiCch}
\alias{ldNcbiCch}
\title{Retrieve cached NCBI query}
\usage{
ldNcbiCch(fnm, args, wd)
}
\arguments{
\item{fnm}{NCBI Entrez function name}

\item{args}{Args used for function}

\item{wd}{Working directory}
}
\description{
Run this function to load cached NCBI queries.
}
