% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-cache.R
\name{ldSqs}
\alias{ldSqs}
\alias{sqs_load}
\title{Load sequences from cache}
\usage{
sqs_load(wd, txid)
}
\arguments{
\item{wd}{Working directory}

\item{txid}{Taxonomic ID, numeric}
}
\description{
Load sequences downloaded by
\code{dwnld} function.
}
