% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache-tools.R
\name{ldSqs}
\alias{ldSqs}
\title{Load sequences from cache}
\usage{
ldSqs(wd = wd, txid = txid)
}
\arguments{
\item{wd}{Working directory}

\item{txid}{Taxonomic ID, numeric}
}
\description{
Load sequences downloaded by
\code{dwnld} function.
}
