% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast-tools.R
\name{mkBlstDB}
\alias{mkBlstDB}
\title{Make a BLAST database}
\usage{
mkBlstDB(sqs, dbfl, wd, verbose = FALSE)
}
\arguments{
\item{sqs}{Sequences}

\item{dbfl}{Outfile for database}

\item{wd}{Working directory}
}
\description{
Generate BLAST database in wd 
for given sequences.
}
