% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage-tools-cluster2.R
\name{mrgClstrs}
\alias{mrgClstrs}
\title{Merge joined clusters}
\usage{
mrgClstrs(jnd_clstrs)
}
\arguments{
\item{jnd_clstrs}{List of joined clusters}
}
\description{
Takes a list of joined clusters and computes
each data slot to create a single merged cluster.
}
