% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage4-tools.R
\name{mrgClstrs}
\alias{mrgClstrs}
\title{Merge joined clusters}
\usage{
mrgClstrs(jnd_clstrs, txdct)
}
\arguments{
\item{jnd_clstrs}{List of joined cluster records}

\item{txdct}{Taxonomic dictionary}
}
\description{
Takes a list of joined clusters and computes
each data slot to create a single merged cluster.
txdct is required for parent look-up.
}
