% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R, R/taxise-tools.R
\name{nDscndnts}
\alias{nDscndnts}
\alias{nDscndnts}
\title{Count descendants}
\usage{
nDscndnts(id, nodes)

nDscndnts(id, nodes)
}
\description{
TODO

Count the number of children
descending from a node in the NCBI taxonomy
dump.
}
\examples{
# TODO
}
