% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi-tools.R
\name{nSqs}
\alias{nSqs}
\title{Count number of sequences for txid}
\usage{
nSqs(txid, direct = FALSE, mx_len = 25000, verbose = FALSE)
}
\arguments{
\item{txid}{Taxonomic ID}

\item{direct}{exp or noexp? T/F}

\item{mx_len}{Maximum number of sequences, numeric.}

\item{verbose}{Verbose? T/F}
}
\description{
Return the number of sequences
associated with a taxonomic ID on NCBI GenBank.
}
