% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-entrez.R
\name{nSqs}
\alias{nSqs}
\title{Count number of sequences for txid}
\usage{
nSqs(txid, ps, drct = FALSE)
}
\arguments{
\item{txid}{Taxonomic ID}

\item{drct}{Node-level only or subtree as well? Default FALSE.}
}
\description{
Return the number of sequences
associated with a taxonomic ID on NCBI GenBank.
}
