% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{parameters_reset}
\alias{parameters_reset}
\title{Change parameters in a working directory}
\usage{
parameters_reset(wd, parameters, values)
}
\arguments{
\item{wd}{Working directory}

\item{parameters}{Parameters to be changed, vector.}

\item{values}{New values for each parameter, vector.}
}
\description{
Reset parameters after running \code{setup()}.
}
\examples{
\dontrun{
  
  # reset and restart aotus run
  setup(wd = 'aotus', txid = 9504)
  run(wd = 'aotus')
  # use ctrl+c or Esc to kill
  # change parameters, e.g. reduce btchsz to 100 if NCBI is blocking
  parameters_reset(wd = 'aotus', parameters = 'btchsz', values = 100)
  # restart
  restart(wd = 'aotus')
}
}
\seealso{
Other run-public: \code{\link{ClstrArc-class}},
  \code{\link{ClstrRec-class}},
  \code{\link{Phylota-class}}, \code{\link{SeqArc-class}},
  \code{\link{SeqRec-class}}, \code{\link{TaxDict-class}},
  \code{\link{TaxRec-class}}, \code{\link{clusters2_run}},
  \code{\link{clusters_run}}, \code{\link{reset}},
  \code{\link{restart}}, \code{\link{run}},
  \code{\link{setup}}, \code{\link{taxise_run}}
}
