% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-phylota.R
\docType{class}
\name{PhyLoTa-class}
\alias{PhyLoTa-class}
\alias{PhyLoTa-method}
\alias{as.character,PhyLoTa-method}
\alias{show,PhyLoTa-method}
\alias{print,PhyLoTa-method}
\alias{str,PhyLoTa-method}
\alias{summary,PhyLoTa-method}
\title{PhyLoTa-class}
\usage{
\S4method{as.character}{PhyLoTa}(x)

\S4method{show}{PhyLoTa}(object)

\S4method{print}{PhyLoTa}(x)

\S4method{str}{PhyLoTa}(object, max.level = 2L, ...)

\S4method{summary}{PhyLoTa}(object)
}
\arguments{
\item{x}{\code{PhyLoTa} object}

\item{object}{\code{PhyLoTa} object}
}
\description{
PhyLoTa table contains sequences and clusters information.
}
\section{Slots}{

\describe{
\item{\code{cids}}{IDs of all clusters}

\item{\code{sids}}{IDs of all sequences}

\item{\code{txids}}{IDs of all taxa}

\item{\code{sqs}}{All sequence records as SqRcrdBx}

\item{\code{cls}}{All cluster records as ClRcrdBx}

\item{\code{txdct}}{Taxonomic dictionary}
}}

\seealso{
\code{\link{genPhyLoTa}}
}
