% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-tools-phylota.R
\name{plotTable}
\alias{plotTable}
\title{Plot table of txid presence by clusters}
\usage{
plotTable(clstrs_obj, clstr_ids = clstrs_obj@clstr_ids,
  txids = clstrs_obj@txids, clstr_names = clstrs_obj@clstr_ids,
  txid_names = clstrs_obj@txids)
}
\arguments{
\item{clstrs_obj}{clstrs_obj}

\item{clstr_ids}{IDs of clusters in table}
}
\description{
TODO
}
