% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-special.R
\name{read_phylota}
\alias{read_phylota}
\title{Generate a Phylota object in R}
\usage{
read_phylota(wd)
}
\arguments{
\item{wd}{Working directory}
}
\value{
Phylota
}
\description{
Creates a Phylota object containing information on
clusters, sequences and taxonomy from the working directory of a
completed pipeline.
}
\examples{
\dontrun{
  
  # Note, this example requires a wd with a completed phylotaR run
  phylota <- read_phylota(wd)
}
}
\seealso{
Other tools-public: \code{\link{calc_mad}},
  \code{\link{calc_wrdfrq}}, \code{\link{drop_by_rank}},
  \code{\link{drop_clstrs}}, \code{\link{drop_sqs}},
  \code{\link{get_clstr_slot}}, \code{\link{get_nsqs}},
  \code{\link{get_ntaxa}}, \code{\link{get_sq_slot}},
  \code{\link{get_stage_times}}, \code{\link{get_tx_slot}},
  \code{\link{get_txids}}, \code{\link{is_txid_in_clstr}},
  \code{\link{is_txid_in_sq}},
  \code{\link{list_clstrrec_slots}},
  \code{\link{list_ncbi_ranks}},
  \code{\link{list_seqrec_slots}},
  \code{\link{list_taxrec_slots}},
  \code{\link{plot_phylota_pa}},
  \code{\link{plot_phylota_treemap}},
  \code{\link{write_sqs}}
}
