% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{reset}
\alias{reset}
\title{Reset a PhyLoTa pipeline run}
\usage{
reset(wd)
}
\arguments{
\item{wd}{Working directory}
}
\description{
Resets the pipeline to a specified stage.
}
\seealso{
\code{\link{run}}
}
