% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster2-tools.R
\name{rnmbrClstrs}
\alias{rnmbrClstrs}
\title{Renumbers cluster IDs}
\usage{
rnmbrClstrs(clstrs)
}
\arguments{
\item{clstrs}{List of clusters}
}
\description{
Returns a list of clusters with
ID determined by the number of sequences
in each cluster.
}
