% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi-tools.R
\name{safeSrch}
\alias{safeSrch}
\title{Safely run rentrez function}
\usage{
safeSrch(func, args, fnm, verbose = FALSE, mx_retry = 1000)
}
\arguments{
\item{func}{rentrez function}

\item{args}{rentrez function arguments, list}

\item{fnm}{rentrez function name}

\item{verbose}{Verbose? T/F}

\item{mx_retry}{Maximum number of attempts}
}
\description{
Safely run a rentrez function.
If the query fails, the function will retry.
}
