% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{setUp}
\alias{setUp}
\title{Set-up parameters}
\usage{
setUp(wd, txid, ncbi_dr = ".", v = FALSE, ...)
}
\arguments{
\item{wd}{Working directory}

\item{txid}{Root taxonomic ID(s), vector or numeric}

\item{ncbi_dr}{Directory to NCBI BLAST tools, default '.'}

\item{...}{Additional parameters}

\item{verbose}{Verbose, T/F}
}
\seealso{
\code{\link{setUp}}
}
