% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache-tools.R
\name{setUpCch}
\alias{setUpCch}
\title{Set-up a cache}
\usage{
setUpCch(wd, prmtrs, ovrwrt = FALSE)
}
\arguments{
\item{wd}{Working directory}

\item{prmtrs}{Paramters list generated by \code{setUpPrmtrs}.}

\item{ovrwrt}{Overwrite existing cache? Default FALSE.}
}
\description{
Creates a cache of parameters in the wd.
}
\details{
Warning: overwriting with this function will
delete the existing cache.
}
