% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-setup.R
\name{setUpNcbiTools}
\alias{setUpNcbiTools}
\title{Ensures NCBI BLAST tools are installed}
\usage{
setUpNcbiTools(d, v, wd)
}
\arguments{
\item{d}{Directory to NCBI BLAST tools}

\item{v}{v, T/F}

\item{wd}{Working directory}
}
\description{
Ensures NCBI BLAST executables are installed on the 
system. Tests version number of BLAST tools.
}
\details{
BLAST tools must be version >=2.0
}
\seealso{
\link{setUpPrmtrs}
}
