% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-setup.R
\name{setUpPrmtrs}
\alias{setUpPrmtrs}
\title{Set Up Parameters}
\usage{
setUpPrmtrs(wd, ncbi_execs, ...)
}
\arguments{
\item{wd}{Working directory}

\item{ncbi_execs}{File directories for NCBI tools, see \code{setUpNcbiTools()}}

\item{...}{Set parameters, see parameters()}
}
\description{
Initiates cache of parameters.
}
\seealso{
\link{parameters}, \link{setUpNcbiTools}
}
