% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-tools.R
\name{setUpPrmtrs}
\alias{setUpPrmtrs}
\title{Return run parameters}
\usage{
setUpPrmtrs(wd, txid, ncbi_execs, tdpth = NULL, mxd = 1e+05, tmout = 100,
  mdlt = 3000, mxsqs = 1000, mxsql = 25000, mxretry = 100,
  mxeval = 1e-10, mncvrg = 51, v = FALSE, ncps = 1)
}
\arguments{
\item{wd}{Working directory}

\item{txid}{Root taxonomic ID(s), vector or numeric}

\item{ncbi_execs}{File directories for NCBI tools, see \code{setUpNcbiTools()}}

\item{tdpth}{NCBI taxdump.tar.gz}

\item{tmout}{Timeout in seconds, numeric.}

\item{mx_dscndnts}{Maximum number of descendants per taxonomic node, numeric}

\item{mdl_thrshld}{Maximum number of sequences per species}

\item{mx_blst_sqs}{Maximum number of sequences to blast in a single run;}

\item{mx_sq_lngth}{Maximum characters in one sequence}

\item{sq_cch_dr}{Directory for sequence cache; will be created if does not exist}

\item{cores}{Number of cores for parellisation}
}
\description{
Provide parameters for a pipeline run
}
\details{
ncbi_execs: users must run \code{setUpNcbiTools()} before running pipeline
mx_blst_sqs: if taxon has more subtree sequences than that,
its children will get clustered.
cores: set to 1 to run sequentially
}
\seealso{
\link{setUpNcbiTools}
}
