% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-tools.R
\name{setUpPrmtrs}
\alias{setUpPrmtrs}
\title{Return run parameters}
\usage{
setUpPrmtrs(wd, txid, ncbi_execs, tdpth = NULL, mx_dscndnts = 10000,
  tmout = 100, mdl_thrshld = 3000, mx_blst_sqs = 10000,
  mx_sq_lngth = 25000, sq_cch_dr = "sequences", verbose = FALSE,
  cores = 1)
}
\arguments{
\item{wd}{Working directory}

\item{txid}{Root taxonomic ID(s), vector or numeric}

\item{ncbi_execs}{File directories for NCBI tools, see \code{setUpNcbiTools()}}

\item{tdpth}{NCBI taxdump.tar.gz}

\item{mx_dscndnts}{Maximum number of descendants per taxonomic node, numeric}

\item{tmout}{Timeout in seconds, numeric.}

\item{mdl_thrshld}{Maximum number of sequences per species}

\item{mx_blst_sqs}{Maximum number of sequences to blast in a single run;}

\item{mx_sq_lngth}{Maximum characters in one sequence}

\item{sq_cch_dr}{Directory for sequence cache; will be created if does not exist}

\item{cores}{Number of cores for parellisation}
}
\description{
Provide parameters for a pipeline run
}
\details{
ncbi_execs: users must run \code{setUpNcbiTools()} before running pipeline
mx_blst_sqs: if taxon has more subtree sequences than that,
its children will get clustered.
cores: set to 1 to run sequentially
}
\seealso{
\link{setUpNcbiTools}
}
