% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-special.R
\name{summary_phylota}
\alias{summary_phylota}
\title{Summarise clusters in PhyLoTa Table}
\usage{
summary_phylota(phylota)
}
\arguments{
\item{phylota}{PhyLoTa object}
}
\description{
Generates a summary data.frame
from all clusters in PhyLoTa object.
}
