% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-cache.R
\name{svBlstCch}
\alias{svBlstCch}
\alias{blastcache_save}
\title{Save BLAST results to cache}
\usage{
blastcache_save(sids, wd, obj)
}
\arguments{
\item{sids}{Sequence IDs}

\item{wd}{Working dir}

\item{obj}{BLAST result}
}
\description{
Run whenever local BLAST runs are made to save
results in cache in case the pipeline is run again.
}
