% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-cache.R
\name{svGIs}
\alias{svGIs}
\alias{sids_save}
\title{Save GIs to cache}
\usage{
sids_save(wd, txid, gis)
}
\arguments{
\item{wd}{Working directory}

\item{txid}{Taxonomic ID, numeric}

\item{gis}{GIs}
}
\description{
Saves GIs downloaded
}
