% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache-tools.R
\name{svSqs}
\alias{svSqs}
\title{Save sequences to cache}
\usage{
svSqs(wd, txid, sqs)
}
\arguments{
\item{wd}{Working directory}

\item{txid}{Taxonomic ID, numeric}

\item{sqdf}{Sequence data.frame}
}
\description{
Saves sequences downloaded \code{getSqs}
and converted to data.frame.
}
\details{
Used within the \code{dwnld} function. Saves
sequence data by txid in cache.
}
