% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes.R
\docType{class}
\name{TxDct-class}
\alias{TxDct-class}
\alias{TxDct-method}
\alias{as.character,TxDct-method}
\alias{show,TxDct-method}
\alias{print,TxDct-method}
\alias{str,TxDct-method}
\alias{summary,TxDct-method}
\title{TxDct-class}
\usage{
\S4method{as.character}{TxDct}(x)

\S4method{show}{TxDct}(object)

\S4method{print}{TxDct}(x)

\S4method{str}{TxDct}(object, max.level = 2L, ...)

\S4method{summary}{TxDct}(object)
}
\arguments{
\item{x}{\code{TxDct} object}

\item{object}{\code{TxDct} object}

\item{max.level}{Maximum level of nesting for str()}

\item{...}{Further arguments for str()}
}
\description{
Taxonomic dictionary contains a taxonomic
tree and NCBI taxonomy data for all taxonomic IDs.
}
\section{Slots}{

\describe{
\item{\code{txids}}{Taxonomic IDs of taxon records}

\item{\code{rcrds}}{Environment of records}

\item{\code{prnt}}{Parent taxonomic ID}

\item{\code{txtr}}{Taxonomic tree}

\item{\code{indx}}{Taxonomic ID index for tree IDs}
}}

