% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-special.R
\name{update_phylota}
\alias{update_phylota}
\title{Update slots}
\usage{
update_phylota(phylota)
}
\arguments{
\item{phylota}{Phylota}
}
\value{
Phylota
}
\description{
After change, run to update slots.
}
\seealso{
Other tools-private: \code{\link{mk_txid_in_sq_mtrx}},
  \code{\link{summary_phylota}}
}
