% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-log.R
\name{warn}
\alias{warn}
\title{Write warning message to log}
\usage{
warn(ps, ...)
}
\arguments{
\item{ps}{Parameters}

\item{...}{Message elements for concatenating}
}
\description{
Inform a user if a potential error has occurred in log.txt.
}
