% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-tools-phylota-special.R
\name{write_phylota}
\alias{write_phylota}
\alias{write_table}
\title{Write out PhyLoTa-like Table}
\usage{
write_table(phylota, outfile)
}
\arguments{
\item{phylota}{Phylota object}

\item{drcty}{Output directory}
}
