% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-tools-phylota-special.R
\name{write_phylota}
\alias{write_phylota}
\title{Write out phylota}
\usage{
write_phylota(phylota, type = c("fasta", "phylota"), cls_names, sqs_names)
}
\arguments{
\item{phylota}{Phylota object}

\item{type}{PhyLoTa or fasta?}

\item{cls_names}{Cluster names}

\item{sqs_names}{Sequence names}
}
\description{
Write out phylota object either
as .fasta or PhyLoTa table.
}
