% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxise-tools.R
\name{writeTax}
\alias{writeTax}
\title{Write out the taxonomic PhyLoTa nodes}
\usage{
writeTax(phylt_nds, td_nms, ps)
}
\arguments{
\item{phylt_nds}{PhyLoTa nodes as generated by \code{genPhylotaNds}}

\item{td_nms}{Taxonomic dump names}

\item{ps}{Parameters}
}
\description{
Writes a .tsv file to disk in the 
PhyLoTa format for all taxonomic information.
}
